% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\name{accessors}
\alias{accessors}
\alias{lsMuModelH0}
\alias{lsMuModelH1}
\alias{lsMuModelConst}
\alias{lsDispModelH0}
\alias{lsDispModelH1}
\alias{lsDispModelConst}
\alias{lsDropModel}
\alias{matCountsProc}
\alias{matWeights}
\alias{scaDFSplinesDisp}
\alias{scaDFSplinesMu}
\alias{strReport}
\alias{vecAllGenes}
\alias{vecConfoundersDisp}
\alias{vecConfoundersMu}
\alias{scaOmega}
\alias{boolFixedPopulation}
\alias{vecH0Pop}
\alias{vecNormConst}
\alias{strVersion}
\alias{dfAnnotationProc}
\alias{lsMuModelH0}
\alias{lsMuModelH1}
\alias{lsMuModelConst}
\alias{lsDispModelH0}
\alias{lsDispModelH1}
\alias{lsDispModelConst}
\alias{lsDropModel}
\alias{matCountsProc}
\alias{matWeights}
\alias{scaDFSplinesDisp}
\alias{scaDFSplinesMu}
\alias{strReport}
\alias{vecAllGenes}
\alias{vecConfoundersDisp}
\alias{vecConfoundersMu}
\alias{scaOmega}
\alias{boolFixedPopulations}
\alias{vecH0Pop}
\alias{vecNormConst}
\alias{strVersion}
\alias{strReport}
\title{LineagePulseObject accession methods}
\usage{
dfAnnotationProc(objLP)

lsMuModelH0(objLP)

lsMuModelH1(objLP)

lsMuModelConst(objLP)

lsDispModelH0(objLP)

lsDispModelH1(objLP)

lsDispModelConst(objLP)

lsDropModel(objLP)

matCountsProc(objLP)

matWeights(objLP)

scaDFSplinesDisp(objLP)

scaDFSplinesMu(objLP)

strReport(objLP)

vecAllGenes(objLP)

vecConfoundersDisp(objLP)

vecConfoundersMu(objLP)

scaOmega(objLP)

boolFixedPopulations(objLP)

vecH0Pop(objLP)

vecNormConst(objLP)

strVersion(objLP)

strReport(objLP)
}
\arguments{
\item{objLP}{(LineagePulse-Object)
A LineagePulse output object to extract from.}
}
\value{
The internal data object specified by the function.
}
\description{
Get internal data of LineagePulse output object.
}
\examples{
   
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 10,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
# get hidden objects within LineagePulse object
dfAnnotationProc <- dfAnnotationProc(objLP)
lsMuModelH0 <- lsMuModelH0(objLP)
lsMuModelH1 <- lsMuModelH1(objLP)
lsMuModelConst <- lsMuModelConst(objLP)
lsDispModelH0 <- lsDispModelH0(objLP)
lsDispModelH1 <- lsDispModelH1(objLP)
lsDispModelConst <- lsDispModelConst(objLP)
lsDropModel <- lsDropModel(objLP)
matCountDataProc <- matCountsProc(objLP)
matWeights <- matWeights(objLP)
scaDFSplinesDisp <- scaDFSplinesDisp(objLP) 
scaDFSplinesMu <- scaDFSplinesMu(objLP) 
strReport <- strReport(objLP) 
vecAllGenes <- vecAllGenes(objLP) 
vecConfoundersDisp <- vecConfoundersDisp(objLP) 
vecConfoundersMu <- vecConfoundersMu(objLP) 
scaOmega <- scaOmega(objLP) 
boolFixedPopulations <- boolFixedPopulations(objLP)
vecH0Pop <- vecH0Pop(objLP) 
vecNormConst <- vecNormConst(objLP) 
strVersion <- strVersion(objLP)

}
\author{
David Sebastian Fischer
}
