% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_calcPostDrop.R
\name{calcPostDrop_Matrix}
\alias{calcPostDrop_Matrix}
\title{Calculate posterior of drop-out}
\usage{
calcPostDrop_Matrix(matCounts, lsMuModel, lsDispModel, lsDropModel,
  vecIDs = NULL)
}
\arguments{
\item{matMu:}{(numeric matrix genes x cells)
Inferred zero inflated negative binomial drop out rates.}

\item{matDisp:}{(vector number of genes) Gene-wise 
negative binomial dispersion coefficients.}

\item{matDrop:}{(numeric matrix genes x cells)
Inferred zero inflated negative binomial drop out rates.}

\item{matboolZero:}{(bool matrix genes x cells)
Whether observation is zero.}

\item{matboolNotZeroObserved:}{(bool matrix genes x cells)
Whether observation is real and non-zero.}
}
\value{
matZ: (numeric matrix genes x cells)
   Posterior probability of observation not being generated 
   by drop-out.
}
\description{
Calculates posterior of observation being a drop-out for a matrix. This posterior
is zero if an observation is non-zero, therefore, the data is read in 
as a boolean matrix indicating zero-observations, the actual values
are not required. Neighbourhood smoothing can be included.
}
\seealso{
Called by \code{fitZINB}. Can be called by user.
}
\author{
David Sebastian Fischer
}
