% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_calcPostDrop.R
\name{calcPostDrop_Vector}
\alias{calcPostDrop_Vector}
\title{Calculate posterior of drop-out}
\usage{
calcPostDrop_Vector(vecMu, vecDisp, vecDrop, vecboolZero, vecboolNotZero)
}
\arguments{
\item{vecMu:}{(numeric vector samples)
Negative binomial mean parameters of samples.}

\item{vecDisp:}{(numeric vector samples)
Negative binomial mean parameters of samples.}

\item{vecDrop:}{(numeric vector samples)
Drop out rates of samples.}

\item{vecboolZero:}{(bool vector samples)
Whether observation is zero.}

\item{vecboolNotZero:}{(bool vector samples)
Whether observation is real and non-zero.}
}
\value{
vecZ:  (numeric vector samples)
   Posterior probability of observation not being generated 
   by drop-out.
}
\description{
Calculates posterior of observation being a drop-out for a vector. This posterior
is zero if an observation is non-zero, therefore, the data is read in 
as a boolean matrix indicating zero-observations, the actual values
are not required. Neighbourhood smoothing can be included. Does not use
parallelisation and takes vectors: As oppose to \code{calcProbNB_Vector}
vectors do not produce format errors and usage within upstream parallelisation
works.
}
\seealso{
Called by \code{fitZINB}. Same function as \code{calcPostDrop_Matrix}
but for vectors (i.e. one gene) as oppose to matrices: no 
parallelisation and no formatting errors.
}
\author{
David Sebastian Fischer
}
