% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\docType{methods}
\name{$,LineagePulseObject-method}
\alias{$,LineagePulseObject-method}
\title{List-like accessor methods for LineagePulseObject: $}
\usage{
\S4method{$}{LineagePulseObject}(x, name)
}
\arguments{
\item{x}{(LineagePulseObject) LineagePulse output object.}

\item{name}{(str) Element from x list to be retrieved.}
}
\value{
Target element from x.
}
\description{
$ accessor function for LineagePulseObject, relies on [[ ]].
Allow usage of LineagePulse ouput object like a list with
respect to the most relevant output:
dfLineagePulseResults and vecDEGenes.
List of all available list-object like accessors:
\link{names,LineagePulseObject-method},
\link{[[,LineagePulseObject,character,missing-method},
\link{$,LineagePulseObject-method}.
}
\examples{
   
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 10,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
head(objLP$dfResults)

}
\author{
David Sebastian Fischer
}
