% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_decompressParameters.R
\name{decompressDispByGeneMM}
\alias{decompressDispByGeneMM}
\title{Compute mean parameter estimate matrix from mean parameter model
for a gene (strDispModel == "MM")}
\usage{
decompressDispByGeneMM(vecDispModel, lsvecBatchModel = NULL,
  lsDispModelGlobal, vecInterval = NULL)
}
\arguments{
\item{vecDispModel}{(numerical vector number of model parameters)
Parameters of dispersion model for given gene.}

\item{lsvecBatchModel}{(list) [Defaul NULL]
List of vectors of batch correction models for dispersion parameter.}

\item{lsDispModelGlobal}{(list)
Object containing meta-data of gene-wise dispersion parameter models.}

\item{vecInterval}{(integer vector length target cells) [Default NULL]
Positions of cells in ordering, for which parameters are to be
computed. Default all cells.}
}
\value{
matDisp (matrix number of cells x number of mixtures)
Dispersion parameter estimates for given gene given the mean model.
}
\description{
Compute mean parameter estimate matrix from mean parameter model
for a gene (strDispModel == "MM")
}
\author{
David Sebastian Fischer
}
