% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_decompressParameters.R
\name{decompressMeansByGene}
\alias{decompressMeansByGene}
\title{Compute mean parameter estimates from mean parameter model for a gene}
\usage{
decompressMeansByGene(vecMuModel, lsvecBatchModel = NULL, lsMuModelGlobal,
  vecInterval = NULL)
}
\arguments{
\item{vecMuModel}{(numerical vector number of model parameters)
Parameters of mean model for given gene.}

\item{lsvecBatchModel}{(list) 
List of vectors of batch correction models for mean parameter.}

\item{lsMuModelGlobal}{(list) 
Object containing meta-data of gene-wise mean parameter models.}

\item{vecInterval}{(integer vector length target cells) [Default NULL]
Positions of cells in ordering, for which parameters are to be 
computed. Default all cells.}
}
\value{
vecMu (numerical vector number of cells)
Mean parameter estimates for given gene given the mean model.
}
\description{
Takes the model type and computes one mean parameter for each cell for one
gene.
}
\seealso{
Called by \code{fitZINB}.
}
\author{
David Sebastian Fischer
}
