% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_decompressParameters.R
\name{decompressMuByGeneMM}
\alias{decompressMuByGeneMM}
\title{Compute mean parameter estimate matrix from mean parameter model
for a gene (strMuModel == "MM")}
\usage{
decompressMuByGeneMM(vecMuModel, lsvecBatchModel = NULL, lsMuModelGlobal,
  vecInterval = NULL)
}
\arguments{
\item{vecMuModel}{(numerical vector number of model parameters)
Parameters of mean model for given gene (means by mixture).}

\item{lsvecBatchModel}{(list) [Defaul NULL]
List of vectors of batch correction models for mean parameter.}

\item{lsMuModelGlobal}{(list)
Object containing meta-data of gene-wise mean parameter models.}

\item{vecInterval}{(integer vector length target cells) [Default NULL]
Positions of cells in ordering, for which parameters are to be
computed. Default all cells.}
}
\value{
matMu (matrix number of cells x number of mixtures)
Mean parameter estimates for given gene given the mean model.
}
\description{
Compute mean parameter estimate matrix from mean parameter model
for a gene (strMuModel == "MM")
}
\author{
David Sebastian Fischer
}
