% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalDropoutModel.R
\name{evalDropoutModel}
\alias{evalDropoutModel}
\title{Compute value of logistic dropout function given with scale boundaries}
\usage{
evalDropoutModel(vecPiModel, vecPiPredictors)
}
\arguments{
\item{vecPiModel}{(numeric vector length linear model) Linear model
for drop-out rate in logit space.}

\item{vecPiPredictors}{(vector length of predictors) Predictors of
the drop-out rate in the linear model. Minimum are a constant
offset and log of the negative binomial mean parameter. 
Other gene-specific predictors can be added.}
}
\value{
scaDropoutRate (scalar) 
   Drop-out rate estimate.
}
\description{
Computes value of logistic function for one observation and uses offset
to correct model. The drop-out magnitude boundary parameter scaOffset is
set in here.
}
\author{
David Sebastian Fischer
}
