% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalDropoutModel.R
\name{evalDropoutModel_comp}
\alias{evalDropoutModel_comp}
\title{Compiled function: evalDropoutModel}
\usage{
evalDropoutModel_comp(vecPiModel, vecPiPredictors)
}
\arguments{
\item{vecPiModel}{(numeric vector length linear model) Linear model
for drop-out rate in logit space.}

\item{vecPiPredictors}{(vector length of predictors) Predictors of
the drop-out rate in the linear model. Minimum are a constant
offset and log of the negative binomial mean parameter. 
Other gene-specific predictors can be added.}
}
\value{
scaDropoutRate (scalar) 
   Drop-out rate estimate.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalDropoutModel}.
}
\seealso{
\link{evalDropoutModel_comp}
}
\author{
David Sebastian Fischer
}
