% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalImpulseModel.R
\name{evalImpulseModel_comp}
\alias{evalImpulseModel_comp}
\title{Compiled function: evalImpulseModel}
\usage{
evalImpulseModel_comp(vecImpulseParam, vecTimepoints)
}
\arguments{
\item{vecImpulseParam}{(numeric vector number of impulse model parameters)
\{beta1, beta2, h0, h1, h2, t1, t2\}
Vector of impulse model parameters.}

\item{vecTimepoints}{(numeric vector length number of time points) 
Time points to be evaluated.}
}
\value{
vecImpulseValue (vec number of vecTimepoints) 
    Model values for given time points.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalImpulseModel}.
}
\author{
David Sebastian Fischer
}
