% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLikZINB.R
\name{evalLikZINB}
\alias{evalLikZINB}
\title{Compute likelihood of zero-inflated negative binomial model
for a vector of counts.}
\usage{
evalLikZINB(vecCounts, vecMu, vecDisp, vecPi, vecboolNotZero, vecboolZero)
}
\arguments{
\item{vecCounts}{(count vector number of samples)
Observed read counts, not observed are NA.}

\item{vecMu}{(vector number of samples) 
Negative binomial mean parameter.}

\item{vecNormConst:}{(numeric vector number of samples) 
Model scaling factors, one per cell.}

\item{vecDisp:}{(scalar vector number of samples) 
Negative binomial dispersion parameters.}

\item{vecPi:}{(probability vector number of samples) 
Drop-out rate estimates.}

\item{vecboolNotZero:}{(bool vector number of samples)
Whether observation is larger than zero.}

\item{vecboolZero:}{(bool vector number of samples)
Whether observation is zero.}
}
\value{
vecLik: (vector number of samples) 
   Likelihood under zero-inflated
	  negative binomial model of each sample.
}
\description{
This liklihood function is appropriate for sequencing data with high drop 
out rate, commonly observed in single cell data (e.g. scRNA-seq). This
is the core function used for every likelihood evaluation 
in LineagePulse, such as maximum likelihood-based estimation. 
It operates on a vector of counts, such as observations of a gene. Note that
for the sake of numerical stability, lower bounds on loglikelihood terms
are implemented.
}
\seealso{
Called directly by likelihood wrappers 
\code{evalLogLikSmoothZINB} and \code{evalLogLikGene}
Called directly by \code{evalLogLikPiZINB},
\code{evalLogLikDispConstMuClustersZINB}
and \code{evalLogLikMuWindowZINB}
Moreover called by \code{plotGene}.
Compiled verison: \link{evalLikZINB_comp}
Loglikelihood: \link{evalLogLikZINB}
}
\author{
David Sebastian Fischer
}
