% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLikZINB.R
\name{evalLikZINB_comp}
\alias{evalLikZINB_comp}
\title{Compiled function: evalLikZINB}
\usage{
evalLikZINB_comp(vecCounts, vecMu, vecDisp, vecPi, vecboolNotZero, vecboolZero)
}
\arguments{
\item{vecCounts}{(count vector number of samples)
Observed read counts, not observed are NA.}

\item{vecMu}{(vector number of samples) 
Negative binomial mean parameter.}

\item{vecNormConst:}{(numeric vector number of samples) 
Model scaling factors, one per cell.}

\item{vecDisp:}{(scalar vector number of samples) 
Negative binomial dispersion parameters.}

\item{vecPi:}{(probability vector number of samples) 
Drop-out rate estimates.}

\item{vecboolNotZero:}{(bool vector number of samples)
Whether observation is larger than zero.}

\item{vecboolZero:}{(bool vector number of samples)
Whether observation is zero.}
}
\value{
scaLik: (scalar) Likelihood under zero-inflated
	  negative binomial model.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalLikZINB}.
}
\seealso{
\link{evalLikZINB}
}
\author{
David Sebastian Fischer
}
