% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitNB_fitMeanDispersion.R
\name{evalLogLikContinuousNB}
\alias{evalLogLikContinuousNB}
\title{Cost function negative binomial model for mean and
dispersion model fitting}
\usage{
evalLogLikContinuousNB(vecTheta, vecCounts, lsMuModelGlobal, lsDispModelGlobal,
  vecTimepoints, vecindTimepointAssign, vecidxNotZero, vecidxZero)
}
\arguments{
\item{vecTheta}{(numeric vector dispersion (1) and impulse parameters (6))
Dispersion and mean parameter estimates.}

\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{lsMuModelGlobal}{(list)
Object containing meta-data of gene-wise mean parameter models.}

\item{lsDispModelGlobal}{(list)
Object containing meta-data of gene-wise dispersion parameter models.}

\item{vecTimepoints}{(numerical vector number of unique time coordinates)
Unique (pseudo)time coordinates of cells.}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to cell in list of sorted
time points. vecTimepoints[vecindTimepointAssign]==vecPseudotime}

\item{vecidxNotZero}{(index vector vector number of cells)
Observation which are larger than zero.}

\item{vecidxZero}{(index vector number of cells)
Observation which are zero.}
}
\value{
scaLogLik (scalar) 
Value of cost function (likelihood) for given gene.
}
\description{
Log likelihood of negative binomial model. 
This function is designed to allow numerical optimisation
of mean and dispersion paramater model on single gen.
}
\author{
David Sebastian Fischer
}
