% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitNB_fitMeanDispersion.R
\name{evalLogLikContinuousNB_comp}
\alias{evalLogLikContinuousNB_comp}
\title{Compiled function: evalLogLikContinuousNB}
\usage{
evalLogLikContinuousNB_comp(vecTheta, vecCounts, lsMuModelGlobal,
  lsDispModelGlobal, vecTimepoints, vecindTimepointAssign, vecidxNotZero,
  vecidxZero)
}
\arguments{
\item{vecTheta}{(numeric vector dispersion (1) and impulse parameters (6)) 
Dispersion and mean parameter estimates.}

\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{lsMuModelGlobal}{(list)
Object containing meta-data of gene-wise mean parameter models.}

\item{lsDispModelGlobal}{(list)
Object containing meta-data of gene-wise dispersion parameter models.}

\item{vecTimepoints}{(numerical vector number of unique time coordinates)
Unique (pseudo)time coordinates of cells.}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to cell in list of sorted
time points. vecTimepoints[vecindTimepointAssign]==vecPseudotime}

\item{vecidxNotZero}{(index vector vector number of cells)
Observation which are larger than zero.}

\item{vecidxZero}{(index vector number of cells)
Observation which are zero.}
}
\value{
scaLogLik (scalar) Value of cost function (likelihood) for given gene.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalLogLikContinuousNB}.
}
\seealso{
\link{evalLogLikContinuousNB}
}
\author{
David Sebastian Fischer
}
