% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_fitMeanDispersion.R
\name{evalLogLikContinuousZINB}
\alias{evalLogLikContinuousZINB}
\title{Cost function zero-inflated negative binomial model for mean and
dispersion model fitting}
\usage{
evalLogLikContinuousZINB(vecTheta, vecCounts, lsMuModelGlobal,
  lsDispModelGlobal, vecTimepoints, vecindTimepointAssign, matDropoutLinModel,
  vecPiConstPredictors, lsDropModelGlobal, vecPiParam = NULL, vecidxNotZero,
  vecidxZero)
}
\arguments{
\item{vecTheta}{(numeric vector dispersion (1) and impulse parameters (6)) 
Dispersion and mean parameter estimates.}

\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{lsMuModelGlobal}{(list)
Object containing meta-data of gene-wise mean parameter models.}

\item{lsDispModelGlobal}{(list)
Object containing meta-data of gene-wise dispersion parameter models.}

\item{vecTimepoints}{(numerical vector number of unique time coordinates)
Unique (pseudo)time coordinates of cells.}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to cell in list of sorted
time points. vecTimepoints[vecindTimepointAssign]==vecPseudotime}

\item{matDropoutLinModel}{(matrix number of cells x number of predictors)
Logistic linear model parameters of the dropout rate 
as a function of the mean and constant gene-wise coefficients.}

\item{vecPiConstPredictors}{(numeric vector constant gene-wise coefficients)
Constant gene-wise coeffiecients, i.e. predictors which are not
the offset and not the mean parameter.}

\item{lsDropModelGlobal}{(list)
Object containing meta-data of cell-wise drop-out parameter models.}

\item{vecPiParam}{(numeric vector number of observations)
Pre-evaluated drop-out model if model is not a function on the mean
parameter to be fit.}

\item{vecidxNotZero}{(index vector vector number of cells)
Observation which are larger than zero.}

\item{vecidxZero}{(index vector number of cells)
Observation which are zero.}
}
\value{
scaLogLik (scalar) Value of cost function (likelihood) for given gene.
}
\description{
Log likelihood of zero inflated  negative binomial model. 
This function is designed to allow numerical optimisation
of mean and dispersion paramater model on single gene given
the drop-out model.
}
\author{
David Sebastian Fischer
}
