% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_cofitMeanDispersion.R
\name{evalLogLikContinuousZINB_comp}
\alias{evalLogLikContinuousZINB_comp}
\title{Compiled function: evalLogLikDispConstMuImpulseZINB}
\usage{
evalLogLikContinuousZINB_comp(vecTheta, vecCounts, lsMuModelGlobal,
  lsDispModelGlobal, vecTimepoints, vecindTimepointAssign, matDropoutLinModel,
  vecPiConstPredictors, lsDropModelGlobal, vecPiParam = NULL, vecboolNotZero,
  vecboolZero)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to cell in list of sorted
time points. vecTimepoints[vecindTimepointAssign]==vecPseudotime}

\item{vecTheta:}{(numeric vector dispersion (1) and impulse parameters (6)) 
Log dispersion and mean parameter estimates.}

\item{vecNormConst:}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecTimepoints:}{(numerical vector number of unique time coordinates)
Unique (pseudo)time coordinates of cells.}

\item{matDropoutLinModel:}{(matrix number of cells x number of predictors)
Logistic linear model parameters of the dropout rate 
as a function of the mean and constant gene-wise coefficients.}

\item{vecPiConstPredictors:}{(numeric vector constant gene-wise coefficients)
Constant gene-wise coeffiecients, i.e. predictors which are not
the offset and not the mean parameter.}

\item{vecboolNotZero:}{(bool vector number of cells)
Whether observation is larger than zero.}

\item{vecboolZero:}{(bool vector number of cells)
Whether observation is zero.}
}
\value{
scaLogLik: (scalar) Value of cost function (likelihood) for given gene.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalLogLikDispConstMuImpulseZINB}.
}
\seealso{
\link{evalLogLikDispConstMuImpulseZINB}
}
\author{
David Sebastian Fischer
}
