% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLik.R
\name{evalLogLikGene}
\alias{evalLogLikGene}
\title{Wrapper for log likelihood of (zero-inflated) negative binomial model
for a vector of counts.}
\usage{
evalLogLikGene(vecCounts, vecMu, vecNormConst, vecDisp, vecPi,
  vecidxNotZero, vecidxZero)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{vecMu}{(numeric vector number of cells) 
Negative binomial mean parameter.}

\item{vecNormConst}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecDisp}{(numeric vector number of cells) 
Negative binomial dispersion parameters.}

\item{vecPi}{(probability vector number of cells) 
Drop-out rate estimates. Set to NULL to evaluate NB
likelihood.}

\item{vecidxNotZero}{(bool vector number of cells)
Whether observation is larger than zero.}

\item{vecidxZero}{(bool vector number of cells)
Whether observation is zero.}
}
\value{
scaLogLik (scalar) Likelihood under zero-inflated
negative binomial model.
}
\description{
Chooses whether to evaluate ZINB or NB model based on input.
}
\author{
David Sebastian Fischer
}
