% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLikNB.R
\name{evalLogLikGeneMM_NB}
\alias{evalLogLikGeneMM_NB}
\title{Wrapper for log likelihood of negative binomial model
for a vector of counts.}
\usage{
evalLogLikGeneMM_NB(vecCounts, matMuParam, vecNormConst, vecDisp, vecidxNotZero,
  vecidxZero, scaNCells)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{matMuParam}{(numeric matrix number of cells x number of mixtures) 
Negative binomial mean parameter matrix with one mean per
cell and per mixture.}

\item{vecNormConst}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecDisp}{(numeric vector number of cells) 
Negative binomial dispersion parameters.}

\item{vecidxNotZero}{(bool vector number of cells)
Whether observation is larger than zero.}

\item{vecidxZero}{(bool vector number of cells)
Whether observation is zero.}

\item{scaNCells}{(scalar)
Number of cells (auxillary parameter, this is length of vecCounts).}
}
\value{
NULL
This will be: scaLogLik (scalar) Likelihood under zero-inflated
negative binomial model.
}
\description{
NOT YET SUPPORTED.
LINEAGEPULSE CODE WILL BE EXTENDED BY MODULAR FUNCTIONALITIES
AND THIS IS ONE INSTANCE OF A PLACEHOLDER USED FOR DEVELOPING.
This likelihood function is a wrapper which choses whether to 
compute smoothed or non-smoothed likelihood. Calling the 
non-smoothed routing avoids the use of a for loop in the case
of no smoothing.
}
\author{
David Sebastian Fischer
}
