% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_fitDropout.R
\name{evalLogLikPiZINB_ManyCells}
\alias{evalLogLikPiZINB_ManyCells}
\title{Cost function zero-inflated negative binomial model for drop-out fitting
for many cells}
\usage{
evalLogLikPiZINB_ManyCells(vecTheta, matCounts, lsMuModel, lsDispModel,
  lsDropModel)
}
\arguments{
\item{vecTheta}{(numeric vector length linear model) 
Parameter estimates for logit linear model for drop-out rate.}

\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}

\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{lsDispModel}{(list)
Object containing description of gene-wise dispersion parameter models.}

\item{lsDropModel}{(list)
Object containing description of cell-wise drop-out parameter models.}
}
\value{
scaLogLik (scalar) Value of cost function:
zero-inflated negative binomial likelihood.
}
\description{
Log likelihood of zero inflated  negative binomial model. 
This function is designed to allow numerical optimisation
of logistic drop-out paramater model on multiple cells given
the negative binomial mean and dispersion parameters.
This function is optimised for memory usage vs evalLogLikPiZINB_SingleCell
at the cost of computation time: The parameter models are not
kept as a gene x cell matrix but as the raw model objects which
are evaluated within the cost function.
}
\seealso{
Called by fitting wrapper: \code{fitPi_ManyCells}.
Calls \code{evalLogLikMatrix}.
Compiled version: \link{evalLogLikPiZINB_ManyCells_comp}.
}
\author{
David Sebastian Fischer
}
