% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_fitDropout.R
\name{evalLogLikPiZINB_ManyCells_comp}
\alias{evalLogLikPiZINB_ManyCells_comp}
\title{Cost function zero-inflated negative binomial model for drop-out fitting}
\usage{
evalLogLikPiZINB_ManyCells_comp(vecTheta, matCounts, lsMuModel, lsDispModel,
  lsDropModel)
}
\arguments{
\item{vecTheta}{(numeric vector length linear model) 
Parameter estimates for logit linear model for drop-out rate.}

\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}

\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{lsDispModel}{(list)
Object containing description of gene-wise dispersion parameter models.}

\item{lsDropModel}{(list)
Object containing description of cell-wise drop-out parameter models.}
}
\value{
scaLogLik (scalar) Value of cost function:
zero-inflated negative binomial likelihood.
}
\description{
Refer to \link{evalLogLikPiZINB_ManyCells}.
}
\seealso{
Compiled version of \link{evalLogLikPiZINB_ManyCells}
}
\author{
David Sebastian Fischer
}
