% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLik.R
\name{evalLogLikZINB}
\alias{evalLogLikZINB}
\title{Compute loglikelihood of zero-inflated negative binomial model
for a vector of counts.}
\usage{
evalLogLikZINB(vecCounts, vecMu, vecDisp, vecPi, vecidxNotZero, vecidxZero)
}
\arguments{
\item{vecCounts}{(count vector number of samples)
Observed read counts, not observed are NA.}

\item{vecMu}{(vector number of samples) 
Negative binomial mean parameter.}

\item{vecDisp}{(scalar vector number of samples) 
Negative binomial dispersion parameters.}

\item{vecPi}{(probability vector number of samples) 
Drop-out rate estimates.}

\item{vecidxNotZero}{(bool vector number of samples)
Whether observation is larger than zero.}

\item{vecidxZero}{(bool vector number of samples)
Whether observation is zero.}
}
\value{
scaLogLik (scalar) Likelihood under zero-inflated
negative binomial model.
}
\description{
This liklihood function is appropriate for sequencing data with high drop 
out rate, commonly observed in single cell data (e.g. scRNA-seq). This
is the core function used for every likelihood evaluation 
in LineagePulse, such as maximum likelihood-based estimation. 
It operates on a vector of counts, such as observations of a gene. 
Note that for the sake of numerical stability, 
lower bounds on loglikelihood terms are implemented.
}
\author{
David Sebastian Fischer
}
