% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLik.R
\name{evalLogLikZINB_comp}
\alias{evalLogLikZINB_comp}
\title{Compiled function: evalLogLikZINB}
\usage{
evalLogLikZINB_comp(vecCounts, vecMu, vecDisp, vecPi, vecidxNotZero, vecidxZero)
}
\arguments{
\item{vecCounts}{(count vector number of samples)
Observed read counts, not observed are NA.}

\item{vecMu}{(vector number of samples) 
Negative binomial mean parameter.}

\item{vecDisp}{(scalar vector number of samples) 
Negative binomial dispersion parameters.}

\item{vecPi}{(probability vector number of samples) 
Drop-out rate estimates.}

\item{vecidxNotZero}{(bool vector number of samples)
Whether observation is larger than zero.}

\item{vecidxZero}{(bool vector number of samples)
Whether observation is zero.}
}
\value{
scaLogLik (scalar) Likelihood under zero-inflated
negative binomial model.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalLogLikZINB}
}
\seealso{
\link{evalLogLikZINB}
}
\author{
David Sebastian Fischer
}
