% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitNB_fitMeanDispersion.R
\name{fitContinuousNB}
\alias{fitContinuousNB}
\title{Optim wrapper for gene-wise models other than mixture model.}
\usage{
fitContinuousNB(vecCounts, vecMuModelGuess, lsvecBatchParamGuessMu,
  lsMuModelGlobal, vecDispGuess, lsvecBatchParamGuessDisp, lsDispModelGlobal)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{vecMuModelGuess}{(numeric vector number of mean model parameters) 
Initialisation for impulse model.}

\item{lsvecBatchParamGuessMu}{(list) 
Object containing initialisation for mean parameter batch correction model.}

\item{lsMuModelGlobal}{(list)
Object containing meta-data of gene-wise mean parameter models.}

\item{vecDispGuess}{(numeric vector number of dispersion model parameters)
Initialisation for dispersion model.}

\item{lsvecBatchParamGuessDisp}{(list) 
Object containing initialisation for 
dispersion parameter batch correction model.}
}
\value{
list
\itemize{
\item vecMuModel (numeric vector number of mu model parameters)
Contains the mean model parameters according to the used model.
\item lsvecBatchModelMu (list) 
Fit of batch correction models for mean parameter to given gene.
\item vecDispModel 
(numeric vector number of dispersion model parameters)
Contains the dispersion model parameters according to the used model.
\item lsvecBatchModelDisp (list) 
Fit of batch correction models for dispersion parameter to given gene.
\item scaConvergence (numeric vector number of genes) 
Convergence status of optim for given gene.
\item scaLL (numeric vector number of genes) 
Likelihood of model fit for given gene.
}
}
\description{
Given a parameter initialisation, this function
performs numerical optimisation using BFGS of the 
likelihood function given the supplied mean and dispersion model.
This is the wrapper that calls optim.
}
\details{
This function performs error handling of the numerical fitting procedure.
This function corrects for the likelihood sensitivity bounds used in the 
cost function.
}
\seealso{
Called once for each gene by \code{fitNBMuDisp}
or within wrapper \code{fitNBImpulse} 
once for each initalisation of each gene.
Calls fitting likelihood functions: 
\code{evalLogLikContinuousNB_comp}.
}
\author{
David Sebastian Fischer
}
