% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_cofitMeanDispersion.R
\name{fitContinuousZINB}
\alias{fitContinuousZINB}
\title{Numerical fitting wrapper for constant dispersion
impulse mean model for single initialisation}
\usage{
fitContinuousZINB(vecCounts, vecMuModelGuess, lsvecBatchParamGuessMu,
  lsMuModelGlobal, vecDispGuess, lsvecBatchParamGuessDisp, lsDispModelGlobal,
  matDropoutLinModel, vecPiConstPredictors, lsDropModelGlobal, vecPiParam)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{scaDispGuess:}{(scalar) Initialisation for dispersion parameter
to be estimated.}

\item{vecParamGuessPeak}{(numeric vector number of parameters [6]) 
Initialisation for impulse model for mean parameters.}

\item{vecNormConst:}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecTimepoints:}{(numerical vector number of unique time coordinates)
Unique (pseudo)time coordinates of cells.}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to cell in list of sorted
time points. vecTimepoints[vecindTimepointAssign]==vecPseudotime}

\item{matDropoutLinModel:}{(matrix number of cells x number of predictors)
Logistic linear model parameters of the dropout rate 
as a function of the mean and constant gene-wise coefficients.}

\item{vecPiConstPredictors:}{(numeric vector constant gene-wise coefficients)
Constant gene-wise coeffiecients, i.e. predictors which are not
the offset and not the mean parameter.}

\item{MAXIT:}{(integer) [Default 1000] Maximum number of 
estimation iterations optim.}

\item{RELTOL:}{(scalar) [Default sqrt(10^(-10))]
  Relative tolerance for optim.
@param trace: (integer) optim control parameter for reporting.
@param REPORT: (integer) optim control parameter for reporting.}
}
\value{
list: (length 4)
   \itemize{
     \item scaDisp: (scalar) Negative binomial dispersion 
       parameter estimate. 
     \item vecImpulseParam: (numeric vector length 6)
       {beta, log(h0), log(h1), log(h2), t1, t2}
       Impulse model parameter estimates.
     \item scaLL: (scalar) Loglikelihood of fit.
     \item scaConvergence: (scalar) Convergence status of optim.
   }
}
\description{
Given a parameter initialisation, this function
performs numerical optimisation using BFGS of the 
likelihood function given the impulse model and a constant
dispersion parameter and returns the fitted (maximum likelihood) model.
This is the wrapper that calls optim.
}
\details{
This function performs error handling of the numerical fitting procedure.
This function corrects for the likelihood sensitivity bounds used in the 
cost function.
}
\seealso{
Called by \code{fitDispConstMuImpulseZINB}. This function
performs optimisation of one impulse model initialisation,
\code{fitDispConstMuImpulseZINB} coordinates the overall fitting
of an impulse model to a gene.
Calls loglikelihood wrapper inside of optim:
\code{evalLogLikDispConstMuImpulseZINB}.
}
\author{
David Sebastian Fischer
}
