% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitNB_fitMeanDispersion.R
\name{fitImpulseNB}
\alias{fitImpulseNB}
\title{Multiple initilalisation wrapper for impulse mean model}
\usage{
fitImpulseNB(vecCounts, vecImpulseParamGuess, lsvecBatchParamGuessMu,
  lsMuModelGlobal, vecDispGuess, lsvecBatchParamGuessDisp, lsDispModelGlobal)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{vecImpulseParamGuess}{(numeric vector number of impulse model parameters) 
Initialisation for impulse model.}

\item{lsvecBatchParamGuessMu}{(list) 
Object containing initialisation for mean parameter batch correction model.}

\item{lsMuModelGlobal}{(list)
Object containing meta-data of gene-wise mean parameter models.}

\item{vecDispGuess}{(numeric vector number of dispersion model parameters)
Initialisation for dispersion model.}

\item{lsvecBatchParamGuessDisp}{(list) 
Object containing initialisation for 
dispersion parameter batch correction model.}

\item{lsDispModelGlobal}{(list)
Object containing meta-data of gene-wise dispersion parameter models.}
}
\value{
list
\itemize{
\item vecMuModel (numeric vector number of mu model parameters)
Contains the mean model parameters according to the used model.
\item lsvecBatchModelMu (list) 
Fit of batch correction models for mean parameter to given gene.
\item vecDispModel (numeric vector number of dispersion model parameters)
Contains the dispersion model parameters according to the used model.
\item lsvecBatchModelDisp (list) 
Fit of batch correction models for dispersion parameter to given gene.
\item scaConvergence (numeric vector number of genes) 
Convergence status of optim for given gene.
\item scaLL (numeric vector number of genes) 
Likelihood of model fit for given gene.
}
}
\description{
Multiple initialisation are tried for the impulse model. Therefore,
this wrapper sits ontop of fitContinuousNB() in the fitting hierarchy
and wraps multiple initialisations at the level of one gene.
}
\seealso{
Called by mean-dispersion co-estimation wrapper 
\code{fitNBMuDisp}.
Calls optimisation wrapper \code{fitContinuousNB} 
for each initialisation.
}
\author{
David Sebastian Fischer
}
