% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_cofitMeanDispersion.R
\name{fitImpulseZINB}
\alias{fitImpulseZINB}
\title{Numerical fitting wrapper for constant dispersion
impulse mean model for multiple initialisation}
\usage{
fitImpulseZINB(vecCounts, vecImpulseParamGuess, lsvecBatchParamGuessMu,
  lsMuModelGlobal, vecDispGuess, lsvecBatchParamGuessDisp, lsDispModelGlobal,
  matDropoutLinModel, vecPiConstPredictors, lsDropModelGlobal, vecPiParam)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{vecImpulseParamGuess}{(numeric vector number of parameters [6]) 
Initialisation for impulse model for mean parameters.}

\item{scaDispGuess:}{(scalar) Initialisation for dispersion parameter
to be estimated.}

\item{lsMuModelGlobal:}{(list) Global variables for mean model,
common to all genes.
\itemize{
  \item strMuModel: (str) {"constant", "impulse", "clusters", 
"windows"} Name of the mean model.
  \item scaNumCells: (scalar) [Default NA] Number of cells
for which model is evaluated. Used for constant model.
  \item vecPseudotime: (numerical vector number of cells)
[Default NA] Pseudotime coordinates of cells. Used for
impulse model.
  \item vecClusterAssign: (integer vector length number of
cells) [Default NA] Index of cluster assigned to each cell.
Used for clusters model.
  \item MAXIT_BFGS_MuDisp: (int) Maximum number of iterations
for BFGS estimation of impulse model with optim (termination criterium).
  \item RELTOL_BFGS_Impulse: (scalar) Relative tolerance of
change in objective function for BFGS estimation of impulse 
model with optim (termination criterium).
}}

\item{vecNormConst:}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{matDropoutLinModel:}{(matrix number of cells x number of predictors)
Logistic linear model parameters of the dropout rate 
as a function of the mean and constant gene-wise coefficients.}

\item{vecPiConstPredictors:}{(numeric vector constant gene-wise coefficients)
Constant gene-wise coeffiecients, i.e. predictors which are not
the offset and not the mean parameter.}
}
\value{
list (length 3)
   \itemize{
     \item scaDisp: (scalar) Negative binomial dispersion 
       parameter estimate. 
     \item vecImpulseParam: (numeric vector length 6)
       {beta, log(h0), log(h1), log(h2), t1, t2}
       Impulse model parameter estimates.
     \item scaConvergence: (scalar) Convergence status of optim.
   }
}
\description{
Computes impulse parameter initialisation for valley
and peak model and uses both and the prior parameter fit
in three separate optimisation runs to obtain the best 
impulse model fit to the data, simultaneous with fitting a 
constant dispersion factor.
}
\seealso{
Called by mean-dispersion co-estimation wrapper \code{fitZINBMuDisp}.
Calls optimisation wrapper \code{fitImpulseOneInitZINB} 
for each initialisation.
}
\author{
David Sebastian Fischer
}
