% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitWrapperLP.R
\name{fitLPModels}
\alias{fitLPModels}
\title{Fit all models necessary for LineagePulse}
\usage{
fitLPModels(objLP, matPiConstPredictors, strMuModel = "constant",
  strDispModelFull = "constant", strDispModelRed = "constant",
  strDropModel = "logistic_ofMu", strDropFitGroup = "PerCell",
  boolEstimateNoiseBasedOnH0 = TRUE, scaMaxEstimationCycles = 20,
  boolVerbose = FALSE, boolSuperVerbose = FALSE)
}
\arguments{
\item{objLP}{(LineagePulseObject)
LineagePulseObject to which null and alternative model are to be fitted.}

\item{matPiConstPredictors}{(numeric matrix genes x number of constant
gene-wise drop-out predictors) Predictors for logistic drop-out 
fit other than offset and mean parameter (i.e. parameters which
are constant for all observations in a gene and externally supplied.)
Is null if no constant predictors are supplied.}

\item{strMuModel}{(str) {"constant", "groups", "MM",
"splines","impulse"}
[Default "impulse"] Model according to which the mean
parameter is fit to each gene as a function of 
pseudotime in the alternative model (H1).}

\item{strDispModelFull}{(str) {"constant", "groups", "splines"}
[Default "constant"] Model according to which dispersion
parameter is fit to each gene as a function of 
pseudotime in the alternative model (H1).}

\item{strDispModelRed}{(str) {"constant", "groups", "splines"}
[Default "constant"] Model according to which dispersion
parameter is fit to each gene as a function of 
pseudotime in the null model (H0).}

\item{strDropModel}{(str) {"logistic_ofMu", "logistic"}
[Default "logistic_ofMu"] Definition of drop-out model.
"logistic_ofMu" - include the fitted mean in the linear model
of the drop-out rate and use offset and matPiConstPredictors.
"logistic" - only use offset and matPiConstPredictors.}

\item{strDropFitGroup}{(str) {"PerCell", "AllCells"}
[Defaul "PerCell"] Definition of groups on cells on which
separate drop-out model parameterisations are fit.
"PerCell" - one parametersiation (fit) per cell
"ForAllCells" - one parametersiation (fit) for all cells}

\item{boolEstimateNoiseBasedOnH0}{(bool) [Default FALSE]
Whether to co-estimate logistic drop-out model with the 
constant null model or with the alternative model. The
co-estimation with the noise model typically extends the
run-time of this model-estimation step strongly. While
the drop-out model is more accurate if estimated based on
a more realistic model expression model (the alternative
model), a trade-off for speed over accuracy can be taken
and the dropout model can be chosen to be estimated based
on the constant null expression model (set to TRUE).}

\item{scaMaxEstimationCycles}{(integer) [Default 20] Maximum number 
of estimation cycles performed in fitZINB(). One cycle
contain one estimation of of each parameter of the 
zero-inflated negative binomial model as coordinate ascent.}

\item{boolVerbose}{(bool) Whether to follow convergence of the 
iterative parameter estimation with one report per cycle.}

\item{boolSuperVerbose}{(bool) Whether to follow convergence of the 
iterative parameter estimation in high detail with local 
convergence flags and step-by-step loglikelihood computation.}
}
\value{
objLP (LineagePulseObject)
LineagePulseObject with models with and fitting reporters added.
}
\description{
Fit alternative H1 and null H0 on both ZINB and NB noise model
to a data set using cycles of coordinate ascent. The algorithm first
fits the either H1 or H0 together with the dropout model by
iterating over cell-wise (dropout models) and gene-wise (negative 
binomial models) parameters. Subsequently, the remaining model (H0 
or H1) is estimated by iterating over zero-inflated negative binomial
mean and dispersion parameter estimation condition on the previously
estimated logistic drop-out model. The NB noise model based models are
estimated in parallel across genes.
}
\seealso{
Called by \code{runLineagePulse}. 
Calls model estimation wrappers: \code{fitContinuousModels}.
}
\author{
David Sebastian Fischer
}
