% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitDropout.R
\name{fitPi_ManyCells}
\alias{fitPi_ManyCells}
\title{Optim wrapper for drop-out model fitting on many cells}
\usage{
fitPi_ManyCells(vecParamGuess, matCounts, lsMuModel, lsDispModel, lsDropModel)
}
\arguments{
\item{vecParamGuess}{(numeric vector length linear model) 
Initial parameter estimates for logit linear model for drop-out rate.}

\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}

\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{lsDispModel}{(list)
Object containing description of gene-wise dispersion parameter models.}

\item{lsDropModel}{(list)
Object containing description of cell-wise drop-out parameter models.}
}
\value{
vecLinModel (numeric vector length linear model) 
Linear model for drop-out rate in logit space for given cell.
}
\description{
This function fits a logistic drop-out model to a set of cells based
on given gene-specific predictors (which enter the linear model).
Parameter estimation of the linear model is performed by maximum
likelihood based on the overall likelihood.
}
\seealso{
Called by drop-out estimation wrapper code in \code{fitPi}.
Calls fitting cost function:
\code{evalLogLikPiZINB_ManyCells_comp}.
}
\author{
David Sebastian Fischer
}
