% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_fitDropout.R
\name{fitPi_SingleCell}
\alias{fitPi_SingleCell}
\title{Optimisation function for drop-out model fitting}
\usage{
fitPi_SingleCell(vecParamGuess, vecCounts, vecMuParam, scaNormConst,
  vecDispParam, matPiAllPredictors, lsDropModelGlobal)
}
\arguments{
\item{vecCounts}{(count vector number of genes)
Observed read counts, not observed are NA.}

\item{vecDropoutLinModel:}{(numeric vector length linear model)
Previous parameterisation of linear model for drop-out 
rate in logit space for given cell.}

\item{matPiConstPredictors:}{(matrix genes x predictors) Predictors of
the drop-out rate in the linear model. Minimum are a constant
offset and log of the negative binomial mean parameter. 
Other gene-specific predictors can be added.}

\item{vecNormConst:}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecidxInterval:}{(integer vector neighbourhood)
Positions of cells within smooting interval (neighbourhood)
of target cell.}

\item{scaTarget:}{(integer) Index of target cell in interval.}
}
\value{
vecLinModel: (numeric vector length linear model) 
   Linear model for drop-out rate in logit space for given cell.
}
\description{
This function fits a logistic drop-out model to a cell based
on given gene-specific predictors (which enter the linear model).
Parameter estimation of the linear model is performed by maximum
likelihood based on the overall likelihood.
}
\seealso{
Called by drop-out estimation wrapper code in \code{fitZINB}.
Calls loglikelihood wrapper inside of optim:
\code{evalLogLikPiZINB}.
}
\author{
David Sebastian Fischer
}
