% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_fitDropout.R
\name{fitZINBPi}
\alias{fitZINBPi}
\title{Global wrapper for fitting of all drop-out models}
\usage{
fitZINBPi(matCounts, lsMuModel, lsDispModel, lsDropModel)
}
\arguments{
\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}

\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{lsDispModel}{(list)
Object containing description of gene-wise dispersion parameter models.}

\item{lsDropModel}{(list)
Object containing description of cell-wise drop-out parameter models.}
}
\value{
vecLinModel (numeric vector length linear model) 
Linear model for drop-out rate in logit space for given cell.
}
\description{
This function fits all drop-out models required on this data set.
}
\seealso{
Called by ZINB fitting wrapper \code{fitZINB}.
Calls optim wrapper \code{fitPi_ManyCells} or \code{fitPi_SingleCell}.
}
\author{
David Sebastian Fischer
}
