% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\name{get_accessors}
\alias{get_accessors}
\alias{get_lsMuModelH1,LineagePulseObject-method}
\alias{get_lsDispModelH1,LineagePulseObject-method}
\alias{get_lsMuModelH0,LineagePulseObject-method}
\alias{get_lsDispModelH0,LineagePulseObject-method}
\alias{get_DropModel,LineagePulseObject-method}
\alias{get_lsFitZINBReporters}
\alias{get_dfAnnotationProc,LineagePulseObject-method}
\alias{get_vecNormConst,LineagePulseObject-method}
\alias{get_scaNProc,LineagePulseObject-method}
\alias{get_scaQThres,LineagePulseObject-method}
\alias{get_strReport,LineagePulseObject-method}
\alias{get_dfAnnotationProc}
\alias{get_lsMuModelH0}
\alias{get_lsMuModelH1}
\alias{get_lsMuModelConst}
\alias{get_lsDispModelH0}
\alias{get_lsDispModelH1}
\alias{get_lsDispModelConst}
\alias{get_matCountsProc}
\alias{get_matWeights}
\alias{get_scaDFSplinesDisp}
\alias{get_scaDFSplinesMu}
\alias{get_strReport}
\alias{get_vecAllGenes}
\alias{get_vecConfoundersDisp}
\alias{get_vecConfoundersMu}
\alias{get_scaOmega}
\alias{get_boolFixedPopulations}
\alias{get_vecH0Pop}
\alias{get_vecNormConst}
\alias{get_strVersion}
\title{LineagePulseObject accession methods}
\usage{
get_dfAnnotationProc(objLP)

get_lsMuModelH0(objLP)

get_lsMuModelH1(objLP)

get_lsMuModelConst(objLP)

get_lsDispModelH0(objLP)

get_lsDispModelH1(objLP)

get_lsDispModelConst(objLP)

get_matCountsProc(objLP)

get_matWeights(objLP)

get_scaDFSplinesDisp(objLP)

get_scaDFSplinesMu(objLP)

get_strReport(objLP)

get_vecAllGenes(objLP)

get_vecConfoundersDisp(objLP)

get_vecConfoundersMu(objLP)

get_scaOmega(objLP)

get_boolFixedPopulations(objLP)

get_vecH0Pop(objLP)

get_vecNormConst(objLP)

get_strVersion(objLP)
}
\arguments{
\item{object}{(objectLineagePulse)  A LineagePulse output object.}
}
\value{
The internal data object specified by the function.
}
\description{
Get internal data of LineagePulse output object.
}
\examples{
   
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 10,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$dfAnnot,
    strMuModel = "impulse")
# get hidden objects within LineagePulse object
dfAnnotationProc <- get_dfAnnotationProc(objLP)
lsMuModelH0 <- get_lsMuModelH0(objLP)
lsMuModelH1 <- get_lsMuModelH1(objLP)
lsMuModelConst <- get_lsMuModelConst(objLP)
lsDispModelH0 <- get_lsDispModelH0(objLP)
lsDispModelH1 <- get_lsDispModelH1(objLP)
lsDispModelConst <- get_lsDispModelConst(objLP)
matCountDataProc <- get_matCountDataProc(objLP)
matWeights <- get_matWeights(objLP)
scaDFSplinesDisp <- get_scaDFSplinesDisp(objLP) 
scaDFSplinesMu <- get_scaDFSplinesMu(objLP) 
strReport <- get_strReport <- function(objLP) 
vecAllGenes <- get_vecAllGenes <- function(objLP) 
vecConfoundersDisp <- get_vecConfoundersDisp <- function(objLP) 
vecConfoundersMu <- get_vecConfoundersMu <- function(objLP) 
scaOmega <- get_scaOmega <- function(objLP) 
boolFixedPopulations <- get_boolFixedPopulations <- function(objLP)
vecH0Pop <- get_vecH0Pop <- function(objLP) 
vecNormConst <- get_vecNormConst <- function(objLP) 
strVersion <- get_strVersion <- function(objLP)

}
\author{
David Sebastian Fischer
}
