% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_getFits.R
\name{getFitsDispersion}
\alias{getFitsDispersion}
\title{Get dispersion model fits}
\usage{
getFitsDispersion(lsDispModel, vecGeneIDs = NULL)
}
\arguments{
\item{lsDispModel}{(list)
Object containing description of gene-wise dispersion parameter models.}

\item{vecGeneIDs}{(vector of strings) 
Gene IDs for which dispersion model fits are to be extracted.}

\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}
}
\value{
matDispParam (numeric matrix genes x cells)
Dispersion parameter fits.
}
\description{
Return dispersion model fits per gene and cell as matrix for chosen model.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 20,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
# Get dispersion parameter fits on alternative model:
# Use H1 model fits.
vecDispersionFits <- getFitsDispersion(
     lsDispModel = objLP@lsDispModelH1,
     vecGeneIDs = rownames(lsSimulatedData$counts)[1])

}
\author{
David Sebastian Fischer
}
