% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_getFits.R
\name{getFitsDropout}
\alias{getFitsDropout}
\title{Get drop-out model fits}
\usage{
getFitsDropout(lsMuModel, lsDropModel, vecGeneIDs = NULL)
}
\arguments{
\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{lsDropModel}{(list)
Object containing description of cell-wise drop-out parameter models.}

\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}

\item{vecIDs}{(vector of strings) [Default NULL]
Gene IDs for which posteriors of drop-out are to be computed.}
}
\value{
matPiParam (numeric matrix genes x cells)
Drop-out rate fits.
}
\description{
Return drop-out model fits per gene and cell as matrix for chosen models.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 20,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
# Get drop-out rate fits on alternative model:
# Use H1 model fits.
vecDropoutFits <- getFitsDropout(
     lsMuModel = objLP@lsMuModelH1,
     lsDropModel = objLP@lsDropModel,
     vecGeneIDs = rownames(lsSimulatedData$counts)[1])

}
\author{
David Sebastian Fischer
}
