% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_getFits.R
\name{getFitsMean}
\alias{getFitsMean}
\title{Get mean model fits}
\usage{
getFitsMean(lsMuModel, vecGeneIDs = NULL)
}
\arguments{
\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{vecGeneIDs}{(vector of strings) 
Gene IDs for which mean model fits are to be extracted.}
}
\value{
(numeric matrix genes x cells)
Mean parameter fits.
}
\description{
Return mean model fits per gene and cell as matrix for chosen model.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 20,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
# Get mean parameter fits on alternative model:
# Use H1 model fits.
vecMeanFits <- getFitsMean(
     lsMuModel = lsMuModelH1(objLP),
     vecGeneIDs = rownames(lsSimulatedData$counts)[1])
#plot(lsSimulatedData$annot$continuous, vecMeanFits)     

}
\author{
David Sebastian Fischer
}
