% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_getNormData.R
\name{getNormData}
\alias{getNormData}
\title{Return depth and batch corrected data}
\usage{
getNormData(matCounts, lsMuModel, boolDepth = TRUE, boolBatch = TRUE)
}
\arguments{
\item{matCounts:}{(numeric matrix genes x cells)
Count data.}

\item{lsMuModel:}{(list) Mean parameter model parameters.}

\item{boolDepth:}{(bool) [Default TRUE] Whether to normalize for sequencing depth.}

\item{boolBatch:}{(bool) [Default TRUE] Whether to normalize for batch.}
}
\value{
matZ: (numeric matrix genes x cells)
   Posterior probability of observation not being generated 
   by drop-out.
}
\description{
The data normalisation is based on the model normalisation used by and inferred
by LineagePulse, e.g. for data visualisation.
}
\seealso{
Called by \code{fitZINB}. Can be called by user.
}
\author{
David Sebastian Fischer
}
