% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_getFits.R
\name{getPostDrop}
\alias{getPostDrop}
\title{Get posteriors of drop-out}
\usage{
getPostDrop(matCounts, lsMuModel, lsDispModel, lsDropModel, vecGeneIDs)
}
\arguments{
\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}

\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{lsDispModel}{(list)
Object containing description of gene-wise dispersion parameter models.}

\item{lsDropModel}{(list)
Object containing description of cell-wise drop-out parameter models.}

\item{vecGeneIDs}{(vector of strings) [Default NULL]
Gene IDs for which posteriors of drop-out are to be computed.}
}
\value{
(numeric matrix genes x cells)
Posterior probability of observation not being generated 
by drop-out.
}
\description{
Return posteriors of drop-out per gene and cell 
as matrix for chosen models.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 20,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
# Get posterior of drop-out on alternative model:
# Use H1 model fits.
vecPosteriorDropoutFits <- getPostDrop(
     matCounts = lsSimulatedData$counts,
     lsMuModel = lsMuModelH1(objLP),
     lsDispModel = lsDispModelH1(objLP),
     lsDropModel = lsDropModel(objLP),
     vecGeneIDs = rownames(lsSimulatedData$counts)[1])

}
\author{
David Sebastian Fischer
}
