% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_initialiseDropModel.R
\name{initDropModel}
\alias{initDropModel}
\title{Initialise drop-out model container object}
\usage{
initDropModel(matCounts, matPiConstPredictors, lsDropModel, strDropModel,
  strDropFitGroup, MAXIT_BFGS_Pi, RELTOL_BFGS_Pi)
}
\arguments{
\item{matCounts}{(matrix genes x cells)
Count data of all cells, unobserved entries are NA.}

\item{matPiConstPredictors}{(numeric matrix genes x number of constant
gene-wise drop-out predictors) [Default NULL]
Predictors for logistic drop-out 
fit other than offset and mean parameter (i.e. parameters which
are constant for all observations in a gene and externally supplied.)
Is null if no constant predictors are supplied.}

\item{lsDropModel}{(list) [Default NULL]}

\item{strDropModel}{(str) {"logistic_ofMu", "logistic", "none"}
[Default "logistic_ofMu"] Definition of drop-out model.
"logistic_ofMu" - include the fitted mean in the linear model
of the drop-out rate and use offset and matPiConstPredictors.
"logistic" - only use offset and matPiConstPredictors.
"none" - negative binomial noise model without zero-inflation.}

\item{strDropFitGroup}{(str) {"PerCell", "AllCells"}
[Defaul "PerCell"] Definition of groups on cells on which
separate drop-out model parameterisations are fit.
"PerCell" - one parametersiation (fit) per cell
"ForAllCells" - one parametersiation (fit) for all cells}

\item{MAXIT_BFGS_Pi}{(sca)
Maximum number of iterations in BFGS estimation of dropout models.
This is a control parameter to optim().}

\item{RELTOL_BFGS_Pi}{(sca) 
Relative tolerance of BFGS estimation of dropout models.
This is a control parameter to optim().}
}
\value{
lsDropModel (list)
Initialisation of drop-out model object.
}
\description{
Either use supplied fits from previous fitting or initialise 
from count data.
}
\seealso{
Called by \code{fitModel}.
}
\author{
David Sebastian Fischer
}
