% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_initialiseMuModel.R
\name{initMuModel}
\alias{initMuModel}
\title{Initialise mean model container object}
\usage{
initMuModel(matCounts, dfAnnotation, vecConfoundersMu, vecNormConst,
  scaDFSplinesMu, matWeights, matMuModelInit, lsmatBatchModelInitMu, strMuModel,
  MAXIT_BFGS_MuDisp, RELTOL_BFGS_MuDisp)
}
\arguments{
\item{matCounts}{(matrix genes x cells)
Count data of all cells, unobserved entries are NA.}

\item{dfAnnotation}{(data frame cells x meta characteristics)
Annotation table which contains meta data on cells.}

\item{vecConfoundersMu}{(vector of strings number of confounders on  mean)
[Default NULL] Confounders to correct for in mu batch
correction model, must be subset of column names of
dfAnnotation which describe condounding variables.}

\item{vecNormConst}{(numeric vector number of cells) 
Model scaling factors, one per cell. These factors linearly 
scale the mean model for evaluation of the loglikelihood.}

\item{scaDFSplinesMu}{(sca) [Default NULL] 
If strMuModel=="splines", the degrees of freedom of the natural
cubic spline to be used as a mean parameter model.}

\item{matWeights}{(numeric matrix cells x mixtures) [Default NULL]
Assignments of cells to mixtures (for strMuModel="MM").}

\item{matMuModelInit}{(numeric matrix genes x mu model parameters)
[Default NULL]
Contains initialisation of mean model parameters 
according to the used model.}

\item{lsmatBatchModelInitMu}{(list) [Default NULL]
Initialisation of batch correction models for mean parameter.}

\item{strMuModel}{(str) {"constant", "groups", "MM",
"splines","impulse"}
[Default "impulse"] Model according to which the mean
parameter is fit to each gene as a function of 
population structure in the alternative model (H1).}

\item{MAXIT_BFGS_MuDisp}{(sca)
Maximum number of iterations in BFGS estimation of Mu/Disp models.
This is a control parameter to optim().}

\item{RELTOL_BFGS_MuDisp}{(sca) 
Relative tolerance of BFGS estimation of Mu/Disp models.
This is a control parameter to optim().}
}
\value{
lsMuModel (list)
Initialisation of mean model object.
}
\description{
Either use supplied fits from previous fitting or initialise 
from count data.
}
\seealso{
Called by \code{fitModel}.
}
\author{
David Sebastian Fischer
}
