% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\docType{class}
\name{LineagePulseObject-class}
\alias{LineagePulseObject-class}
\title{Container class for LineagePulse output}
\description{
LineagePulse output and intermediate results such as model fits.
}
\section{Slots}{

\describe{
\item{\code{dfDEAnalysis}}{(data frame samples x reported characteristics) 
Summary of fitting procedure and 
differential expression results for each gene.
\itemize{
  \item Gene: Gene ID.
  \item p: P-value for differential expression.
  \item padj: Benjamini-Hochberg false-discovery rate corrected p-value
  for differential expression analysis.
  \item loglik_full: Loglikelihood of full model.
  \item loglik_red: Loglikelihood of reduced model.
  \item df_full: Degrees of freedom of full model.
  \item df_red: Degrees of freedom of reduced model
  \item mean: Inferred mean parameter of constant model of first batch.
  From combined samples in case-ctrl. 
  \item allZero (bool) Whether there were no observed non-zero observations of this gene.
  If TRUE, fitting and DE analsysis were skipped and entry is NA.
}
Entries only present in case-only DE analysis:
\itemize{
  \item converge_impulse: Convergence status of optim for 
  impulse model fit (full model).
  \item converge_const: Convergence status of optim for 
  constant model fit (reduced model).
}
Entries only present in mixture model DE analysis:
\itemize{
  \item converge_mixture: Convergence status of optim for 
  mixture model fit (full model).
  \item converge_const: Convergence status of optim for 
  constant model fit (reduced model).
}
Entries only present if boolIdentifyTransients is TRUE:
\itemize{
  \item converge_sigmoid: Convergence status of optim for 
  sigmoid model fit to samples of case condition.
  \item impulseTOsigmoid_p: P-value of loglikelihood ratio test
  impulse model fit versus sigmoidal model on samples of case condition.
  \item impulseTOsigmoid_padj: Benjamini-Hochberg 
  false-discovery rate corrected p-value of loglikelihood ratio test
  impulse model fit versus sigmoid model on samples of case condition.
  \item sigmoidTOconst_p: P-value of loglikelihood ratio test
  sigmoidal model fit versus constant model on samples of case condition.
  \item sigmoidTOconst_padj: Benjamini-Hochberg 
  false-discovery rate corrected p-value of loglikelihood ratio test
  sigmoidal model fit versus constant model on samples of case condition.
  \item isTransient (bool) Whether gene is transiently
  activated or deactivated and differentially expressed.
  \item isMonotonous (bool) Whether gene is not transiently
  activated or deactivated and differentially expressed. This scenario
  corresponds to a montonous expression level increase or decrease.
}}

\item{\code{strReport}}{(str) LineagePulse stdout report.}
}}

\author{
David Sebastian Fischer
}
