% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\docType{methods}
\name{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{get_lsMuModelH1,LineagePulseObject-method}
\alias{get_lsDispModelH1,LineagePulseObject-method}
\alias{get_lsMuModelH0,LineagePulseObject-method}
\alias{get_lsDispModelH0,LineagePulseObject-method}
\alias{get_DropModel,LineagePulseObject-method}
\alias{get_lsFitZINBReporters}
\alias{get_dfAnnotationProc,LineagePulseObject-method}
\alias{get_vecNormConst,LineagePulseObject-method}
\alias{get_scaNProc,LineagePulseObject-method}
\alias{get_scaQThres,LineagePulseObject-method}
\alias{get_strReport,LineagePulseObject-method}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\alias{LineagePulseObject_Accessors}
\title{LineagePulseObject accession methods}
\usage{
\S4method{get_lsMuModelH1}{LineagePulseObject}(object)

\S4method{get_lsDispModelH1}{LineagePulseObject}(object)

\S4method{get_lsMuModelH0}{LineagePulseObject}(object)

\S4method{get_lsDispModelH0}{LineagePulseObject}(object)

\S4method{get_DropModel}{LineagePulseObject}(object)

\S4method{get_lsFitZINBReporters}{LineagePulseObject}(object)

\S4method{get_dfAnnotationProc}{LineagePulseObject}(object)

\S4method{get_vecNormConst}{LineagePulseObject}(object)

\S4method{get_scaNProc}{LineagePulseObject}(object)

\S4method{get_scaQThres}{LineagePulseObject}(object)

\S4method{get_strReport}{LineagePulseObject}(object)
}
\arguments{
\item{object}{(objectLineagePulse)  A LineagePulse output object.}
}
\value{
The internal data object specified by the function.
}
\description{
Get internal data of LineagePulse output object.
}
