% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_plotGene.R
\name{plotCellDensity}
\alias{plotCellDensity}
\title{Plot density of cells in pseudotime}
\usage{
plotCellDensity(objLP, bwDensity = NULL)
}
\arguments{
\item{objLP}{(LineagePulseObject) LineagePulseObject to base plot on.}
}
\value{
gGenePlot: (ggplot object)
ggplot2 kernel density estimator plot.
}
\description{
Plot density of cells in pseudotime
}
\author{
David Sebastian Fischer
}
