% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_plotComparativeECDF.R
\name{plotComparativeECDF}
\alias{plotComparativeECDF}
\title{Plot comparative q-value ECDF}
\usage{
plotComparativeECDF(lsObjectLineagePulse, filePDF = NULL, strRunLabels,
  strTitle = "Comparative Q-value ECDF")
}
\arguments{
\item{lsObjectLineagePulse:}{(list of LineagePulseObjects)
List of LineagePulseObjects with DE results table to 
base ECDF on.}

\item{filePDF:}{(directory/filename)
File to which plot is saved (pdf).}

\item{strRunLabels:}{(string vector arbitrary length)
Names of runs to be plotted. Correspond
to directories supplied in lsDirOutLineagePulse.}

\item{strTitle:}{(string) Title of plot.}
}
\description{
Loads DE results data frame from list of given LineagePulse ouput folders
and plots ECDF for each output into single plot.
}
\seealso{
Auxillary method not called by LineagePulse wrapper.
Called separately by user.
}
\author{
David Sebastian Fischer
}
