% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_plotGene.R
\name{plotGene}
\alias{plotGene}
\title{Plot counts and model for one gene}
\usage{
plotGene(objLP, strGeneID, vecReferenceMuParam = NULL,
  strTitleSuffix = NULL, boolLogPlot = TRUE, boolColourByDropout = TRUE,
  boolH1NormCounts = FALSE, boolLineageContour = FALSE, bwDensity = NULL)
}
\arguments{
\item{objLP}{(LineagePulseObject) LineagePulseObject
base plot on.}

\item{strGeneID}{(str) Name of gene, used for title of plot.}

\item{vecReferenceMuParam}{(numeric vector length number of cells)
[Default NULL] Reference mean trajectory which can be plotted}

\item{strTitleSuffix}{(str) String to be added to title.}

\item{boolColourByDropout}{(bool) Whether to colour scatter
plot by posterior of drop-out.}

\item{boolLineageContour}{(bool) [Default FALSE]
Whether to the "lineage contour" lines to the scatter plot.}

\item{bwDensity}{(bandwith: numeric or string) [Default NULL]
Bandwith to be used to kernel density smooting
of cell density in pseudotime (used if boolExpectedDistrPanel=TRUE).
If not set, defaults to stats:density() default.}
}
\value{
gGenePlot: (ggplot object)
   Model rajectories and scatter plot for given gene.
}
\description{
Plot counts and model for one gene as a scatter plot with model trajectories:
Both as counts or log10 counts vs pseudotime.
Dropout rates can be given and are visualised as the colour of the observation
points.
A reference model trajectory can be added.
Contour plots: A contour plot aids the interpretation of the model trajectory
in context of the observed data if local overplotting occurs because of a
non-uniform distribution of cells in pseudotime (a very common scenario).
In such a case, one might mistake cells with high counts from a pseudotime
interval with high cellular intensity as a indication of increased expression
in this interval. Instead, one can explain such local maxima based on the 
simple phenomenom that more samples were taken from the underyling distribution
in that pseudotime interval which increases the probability of sampling
far in the tail of the distribution. To migitate this visually, we added the
option to add "contours": A contour is an alternative to a confidence interval
which includes information on the local sampling density. The contour is specific
to a number of cells n. It is an expression trajectory in pseudotime which lies
at the line above which n cells are expected within a local pseudotime bin given
its sampling density and the H1 non-constant expression model. Contours therefore
quantify the increase in maximum observed counts in pseudotime intervals due to
sampling density, which is not due to the expression model. Note that this effect
is corrected for in the model fitting and that the contour plots just aids visual
interpretation of the fit to the data!
}
\author{
David Sebastian Fischer
}
