% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_runDEAnalysis.R
\name{runDEAnalysis}
\alias{runDEAnalysis}
\title{Differential expression analysis}
\usage{
runDEAnalysis(objLP)
}
\arguments{
\item{:}{objLP: (LineagePulseObject)
LineagePulseObject with fitted null and alternative models.}
}
\value{
objLP: (LineagePulseObject)
LineagePulseObject with analysis summary table (dfResults)
added.
}
\description{
Performs differential expression analysis based on previously estimated 
null and alternative models.
(I) Compute loglikelihood of data under null H0 and alternative H1 model.
(II) Differential expression analysis as loglikelihood ratio test.
}
\seealso{
Called by \code{runLineagePulse}.
}
\author{
David Sebastian Fischer
}
