% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_simulateDataSet.R
\name{simulateContinuousDataSet}
\alias{simulateContinuousDataSet}
\title{Simulate a data set for LinagePulse
Simulates a data set with genes with constant and impulse
expression traces. Expression strength and variation in impulse
like traces are parameterised and random. All temporary files
are saved into dirOutSimulation and only the objects necessary
for running LineagePulse (the count matrix and the pseudotime
vector are returned). The remaining objects representing hidden
parameters can be used to evaluate parameter estimates. Cells
are distributed uniformly in pseudotime.}
\usage{
simulateContinuousDataSet(scaNCells, scaNConst, scaNLin, scaNImp,
  scaMumax = 100, scaSDMuAmplitude = 1, vecNormConstExternal = NULL,
  vecDispExternal = NULL, vecGeneWiseDropoutRates = NULL,
  matDropoutModelExternal = NULL)
}
\arguments{
\item{scaNCells}{(scalar) Number of cells in data set.}

\item{scaNConst}{(scalar) Number of constant expression profiles (genes) in data set.}

\item{scaNLin}{(scalar) Number of linear expression profiles (genes) in data set.}

\item{scaNImp}{(scalar) Number of impulse model expression profiles (genes) in data set.}

\item{scaMumax}{(scalar) [Default 1000]
Maximum expression mean parameter to be used.}

\item{scaSDMuAmplitude}{(scalar) [Default 1]
Standard deviation of normal distribution form which the 
amplitude change within an impulse trace is drawn.}

\item{vecNormConstExternal}{(numeric vector number of cells)
[Default NULL]
Size factors for data set. Size factors are set to 1 if this is
not specified (NULL).}

\item{vecDispExternal}{(numeric vector number of genes)
[Default NULL]
Dispersion parameters per gene supplied by user.s}

\item{vecGeneWiseDropoutRates}{(numeric vector number of cells) 
[Default NULL] One drop-out rate per gene.}

\item{matDropoutModelExternal}{(numeric matrix cells x 2) 
[Default NULL] External drop-out model, has to have
one row for each simulated cell.}
}
\value{
list (length 2)
\itemize{
\item pseudotime (numerical vector length number of cells)
Pseudotime coordinates (1D) of cells. One scalar per cell.
\item counts (matrix genes x cells)
Sampled count data of all cells after drop-out.
}
}
\description{
Set either vecGeneWiseDropoutRates or matDropoutModelExternal.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 100,
    scaNConst = 10,
    scaNLin = 10,
    scaNImp = 10,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 30),
    vecGeneWiseDropoutRates = rep(0.1, 30))
plot(lsSimulatedData$annot$pseudotime, lsSimulatedData$counts[1,])

}
\seealso{
Called by separately by user.
}
\author{
David Sebastian Fischer
}
