% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_simulateDataSet.R
\name{simulateDataSet}
\alias{simulateDataSet}
\title{Simulate a data set for LinagePulse
Simulates a data set with genes with constant and impulse
expression traces. Expression strength and variation in impulse
like traces are parameterised and random. All temporary files
are saved into dirOutSimulation and only the objects necessary
for running LineagePulse (the count matrix and the pseudotime
vector are returned). The remaining objects representing hidden
parameters can be used to evaluate parameter estimates. Cells
are distributed uniformly in pseudotime.}
\usage{
simulateDataSet(scaNCells, scaNConst, scaNImp, scaPTmax = 100,
  scaMumax = 1000, scaSDImpulseAmplitude = 1, vecNormConstExternal = NULL,
  matDropoutModelExternal = NULL, strDropModel = "logistic_ofMu",
  strDropFitGroup = "PerCell", dirOutSimulation = NULL)
}
\arguments{
\item{scaNCells:}{(scalar) Number of cells in data set.}

\item{scaNConst:}{(scalar) Number of constant genes in data set.}

\item{scaNImp:}{(scalar) Number of impulse distributed genes in data set.}

\item{scaPTmax:}{(scalar) [Default 100]
Maximum pseudotime coordinate of all cells. This doesnt
really matter as cells are uniformly distributed in the 
current implementation.}

\item{scaMumax:}{(scalar) [Default 1000]
Maximum expression mean parameter to be used.}

\item{scaSDImpulseAmplitude:}{(scalar) [Default 1]
Standard deviation of normal distribution form which the 
amplitude change within an impulse trace is drawn.}

\item{vecNormConstExternal:}{(numeric vector number of cells)
[Default NULL]
Size factors for data set. Size factors are set to 1 if this is
not specified (NULL).}

\item{matDropoutModelExternal:}{(numeric matrix cells x 2) 
[Default NULL] External drop-out model, has to have
one row for each simulated cell.}

\item{dirOutSimulation:}{(str directory)
Directory to which simulated parameter objects are 
saved to.}
}
\value{
list: (length 2)
\itemize{
   \item vecPT: (numerical vector length number of cells)
   Pseudotime coordinates (1D) of cells: One scalar per cell.
   \item matSampledCountsObserved: (matrix genes x cells)
   Sampled count data of all cells after drop-out.
}
}
\description{
Simulate a data set for LinagePulse
Simulates a data set with genes with constant and impulse
expression traces. Expression strength and variation in impulse
like traces are parameterised and random. All temporary files
are saved into dirOutSimulation and only the objects necessary
for running LineagePulse (the count matrix and the pseudotime
vector are returned). The remaining objects representing hidden
parameters can be used to evaluate parameter estimates. Cells
are distributed uniformly in pseudotime.
}
\seealso{
Called by separately by user.
}
\author{
David Sebastian Fischer
}
