% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\name{strReport<-}
\alias{strReport<-}
\title{LineagePulseObject strReport setter}
\usage{
strReport(objLP) <- value
}
\arguments{
\item{objLP}{(LineagePulse-Object)
A LineagePulse output object to write into.}

\item{value}{(str) The string to add to strReport.}
}
\value{
The LineagePulse output object with updated strReport.
}
\description{
Add string to strReport LineagePulse-Object.
}
\author{
David Sebastian Fischer
}
