% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\docType{methods}
\name{[[,LineagePulseObject,character,missing-method}
\alias{[[,LineagePulseObject,character,missing-method}
\title{List-like accessor methods for LineagePulseObject: names()}
\usage{
\S4method{[[}{LineagePulseObject,character,missing}(x, i, j, ...)
}
\arguments{
\item{x}{(LineagePulseObject) LineagePulse output object.}

\item{i}{(str) Element from x list to be retrieved.}

\item{j}{() Ignored argument to generic.}

\item{...}{() Ignored argument to generic.}
}
\value{
Target element from x.
}
\description{
names() function for LineagePulseObject.
Allow usage of LineagePulse ouput object like a list with
respect to the most relevant output:
dfLineagePulseResults and vecDEGenes.
List of all available list-object like accessors:
\link{names,LineagePulseObject-method},
\link{[[,LineagePulseObject,character,missing-method},
\link{$,LineagePulseObject-method}.
}
\examples{
   
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 10,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
head(objLP[["dfResults"]])

}
\author{
David Sebastian Fischer
}
