% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_runHypothesisTests.R
\name{testDropout}
\alias{testDropout}
\title{Test for existance of drop-out with log-likelihood ratio test}
\usage{
testDropout(objLP)
}
\arguments{
\item{objLP}{(LineagePulseObject)
LineagePulseObject with fitted null and alternative models.}
}
\value{
(data frame)
Summary of hypothesis test
\itemize{
\item Gene: Gene ID.
\item p: P-value for existance of drop-out in data set. If high, the
data can be explained with models based on NB noise and zero-inflation
is not necessary: The null hypothesis of no zero-inflation cannot be
rejected.
\item loglik_zinb: Loglikelihood of full model with ZINB noise (all genes).
\item loglik_nb: Loglikelihood of reduced model with NB noise (all genes).
\item df_full: Degrees of freedom of full model with ZINB noise (all genes).
\item df_red: Degrees of freedom of reduced model with NB noise (all genes).
}
}
\description{
Performs one test for entire data set.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 100,
    scaNConst = 10,
    scaNLin = 10,
    scaNImp = 10,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 30),
    vecGeneWiseDropoutRates = rep(0.1, 30))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
testDropout(objLP)

}
\seealso{
Called by user.
}
\author{
David Sebastian Fischer
}
