% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\name{writeReport}
\alias{writeReport}
\title{Print LineagePulse report}
\usage{
writeReport(object, file = "")
}
\arguments{
\item{object}{(LineagePulseObject) Output object of LineagePulse.}

\item{file}{(file) [DEFAULT ""] File to print report to. Default is stdout.}
}
\description{
Print LineagePulse report string to .txt file or stdout.
}
\examples{
   
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 10,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$dfAnnot,
    strMuModel = "impulse")
writeReport(objLP, file="")
  
}
\author{
David Sebastian Fischer
}
