% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_calcNormConst.R
\name{calcNormConst}
\alias{calcNormConst}
\title{Compute size factors for a LineagePulse-object}
\usage{
calcNormConst(objLP, vecNormConstExternal)
}
\arguments{
\item{objLP}{(LineagePulse-object)
Object to fit normalization constants on.}

\item{vecNormConstExternal}{(numeric vector number of cells) 
Model scaling factors supplied by user, one per cell.}
}
\value{
objLP: (LineagePulse-object)
Object with fit normalization constants.
}
\description{
Either use externally supplied normalisation constants or set these to one.
}
\seealso{
Called by \code{runLineagePulse}.
}
\author{
David Sebastian Fischer
}
