% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_calcNormConst.R
\name{calcNormConst}
\alias{calcNormConst}
\title{Compute size factors for a scRNA-seq dataset}
\usage{
calcNormConst(objLP, vecNormConstExternal)
}
\arguments{
\item{matCountsProc:}{(matrix genes x samples)
Count data.}

\item{vecNormConstExternal:}{(numeric vector number of cells) 
Model scaling factors, one per cell. These factors will linearly 
scale the mean model for evaluation of the loglikelihood. 
Must be named according to the column names of matCounts.
Supplied by user.}
}
\value{
vecNormConst: (numeric vector number of cells) 
   Model scaling factors, one per cell.
}
\description{
Either use externally supplied normalisation constants or set these to one.
}
\details{
Size factors scale the negative binomial likelihood
model of a gene to the sequencing depth of each sample.
As of now, no stable normalisation constants are known for
scRNAseq data and they are set to 1 here.
}
\seealso{
Called by \code{runLineagePulse}.
}
\author{
David Sebastian Fischer
}
