% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_calcPostDrop.R
\name{calcPostDrop_Matrix}
\alias{calcPostDrop_Matrix}
\title{Calculate posterior of drop-out}
\usage{
calcPostDrop_Matrix(matCounts, lsMuModel, lsDispModel, lsDropModel,
  vecIDs = NULL)
}
\arguments{
\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}

\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{lsDispModel}{(list)
Object containing description of gene-wise dispersion parameter models.}

\item{lsDropModel}{(list)
Object containing description of cell-wise drop-out parameter models.}

\item{vecIDs}{(vector of strings) [Default NULL]
Gene IDs for which posteriors of drop-out are to be computed.}
}
\value{
matZ (numeric matrix genes x cells)
Posterior probability of observation not being generated 
by drop-out.
}
\description{
Calculates posterior of observation being a drop-out for a matrix.
}
\seealso{
Called by \code{plotGene}.
}
\author{
David Sebastian Fischer
}
