% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_decompressParameters.R
\name{decompressMeansByGene}
\alias{decompressMeansByGene}
\title{Compute mean parameter estimates from mean parameter model for a gene}
\usage{
decompressMeansByGene(vecMuModel, lsvecBatchModel = NULL, lsMuModelGlobal,
  vecInterval = NULL)
}
\arguments{
\item{vecMuModel:}{(numerical vector number of model parameters)
Parameters of mean model for given gene.}

\item{lsMuModelGlobal:}{(list) Global variables for mean model,
common to all genes.
\itemize{
  \item strMuModel: (str) {"constant", "impulse", "clusters", 
      "windows"} Name of the mean model
  \item scaNumCells: (scalar) [Default NA] Number of cells
      for which model is evaluated. Used for constant model.
  \item vecPseudotime: (numerical vector number of cells)
      [Default NA] Pseudotime coordinates of cells. Used for
      impulse model.
  \item vecindClusterAssign: (integer vector length number of
      cells) [Default NA] Index of cluster assigned to each cell.
      Used for clusters model.
}}

\item{vecInterval:}{(integer vector length target cells) [Default NULL]
Positions of cells in ordering, for which parameters are to be 
computed. Default: all cells.}
}
\value{
vecMu: (numerical vector number of cells)
   Mean parameter estimates for given gene given the mean model.
}
\description{
Takes the model type and computes one mean parameter for each cell for one
gene.
}
\seealso{
Called by \code{fitZINB}.
}
\author{
David Sebastian Fischer
}
