% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_cofitMeanDispersion.R
\name{evalLogLikContinuousZINB}
\alias{evalLogLikContinuousZINB}
\title{Cost function zero-inflated negative binomial model for dispersion
and mean co-estimation under constant dispersion and constant mean model}
\usage{
evalLogLikContinuousZINB(vecTheta, vecCounts, lsMuModelGlobal,
  lsDispModelGlobal, vecTimepoints, vecindTimepointAssign, matDropoutLinModel,
  vecPiConstPredictors, lsDropModelGlobal, vecPiParam = NULL, vecboolNotZero,
  vecboolZero)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to cell in list of sorted
time points. vecTimepoints[vecindTimepointAssign]==vecPseudotime}

\item{vecTheta:}{(numeric vector dispersion (1) and impulse parameters (6)) 
Log dispersion and mean parameter estimates.}

\item{vecNormConst:}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecTimepoints:}{(numerical vector number of unique time coordinates)
Unique (pseudo)time coordinates of cells.}

\item{matDropoutLinModel:}{(matrix number of cells x number of predictors)
Logistic linear model parameters of the dropout rate 
as a function of the mean and constant gene-wise coefficients.}

\item{vecPiConstPredictors:}{(numeric vector constant gene-wise coefficients)
Constant gene-wise coeffiecients, i.e. predictors which are not
the offset and not the mean parameter.}

\item{vecboolNotZero:}{(bool vector number of cells)
Whether observation is larger than zero.}

\item{vecboolZero:}{(bool vector number of cells)
Whether observation is zero.}
}
\value{
scaLogLik: (scalar) Value of cost function (likelihood) for given gene.
}
\description{
Log likelihood of zero inflated  negative binomial model. 
This function is designed to allow simultaneous numerical optimisation
of negative binomial dispersionmean paramater on single gene given
the drop-out rate/model. The dispersion parameter is modelled as a constant
and the mean parameter is modelled by the impulse model for the given gene.
}
\details{
The impulse model amplitude parameters are fit in log space and are
therefore fit as positive scalars. The cost function is insensitive to the
amplitude parameters shrinking beyond a numerical threshold to zero which 
may cause numerical errors.
The dispersion parameter is fit in log space and is therefore fit
as a positive scalar. The cost function is insensitive to the
dispersion factor shrinking beyond a numerical threshold to zero
and to growth above a threshold to avoid shrinkage of the 
dispersion factor to zero/ expansion to infinity.
}
\seealso{
Called by fitting wrapper:
\code{fitImpulseOneInitZINB}.
Calls \code{evalImpulseModel} and \code{evalLogLikGene}.
Compiled function: \link{evalLogLikDispConstMuImpulseZINB_comp}.
}
\author{
David Sebastian Fischer
}
