% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLikZINB.R
\name{evalLogLikGene}
\alias{evalLogLikGene}
\title{Wrapper for log likelihood of zero-inflated negative binomial model
for a vector of counts.}
\usage{
evalLogLikGene(vecCounts, vecMu, vecNormConst, vecDisp, vecPi, vecboolNotZero,
  vecboolZero)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{vecMu}{(numeric vector number of cells) 
Negative binomial mean parameter.}

\item{vecNormConst:}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecDisp:}{(numeric vector number of cells) 
Negative binomial dispersion parameters.}

\item{vecPi:}{(probability vector number of cells) 
Drop-out rate estimates.}

\item{vecboolNotZero:}{(bool vector number of cells)
Whether observation is larger than zero.}

\item{vecboolZero:}{(bool vector number of cells)
Whether observation is zero.}
}
\value{
scaLogLik: (scalar) Likelihood under zero-inflated
	  negative binomial model.
}
\description{
This likelihood function is a wrapper which choses whether to 
compute smoothed or non-smoothed likelihood. Calling the 
non-smoothed routing avoids the use of a for loop in the case
of no smoothing.
}
\author{
David Sebastian Fischer
}
