% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLik.R
\name{evalLogLikGeneMM}
\alias{evalLogLikGeneMM}
\title{Wrapper for log likelihood of zero-inflated negative binomial model
for a vector of counts.}
\usage{
evalLogLikGeneMM(vecCounts, matMuParam, vecNormConst, vecDisp, vecPi,
  vecidxNotZero, vecidxZero, scaNCells)
}
\arguments{
\item{vecCounts}{(count vector number of cells)
Observed read counts, not observed are NA.}

\item{matMuParam}{(numeric matrix number of cells x number of mixtures) 
Negative binomial mean parameter matrix with one mean per
cell and per mixture.}

\item{vecNormConst}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecDisp}{(numeric vector number of cells) 
Negative binomial dispersion parameters.}

\item{vecPi}{(probability vector number of cells) 
Drop-out rate estimates. Set to NULL to evaluate NB model.}

\item{vecidxNotZero}{(bool vector number of cells)
Whether observation is larger than zero.}

\item{vecidxZero}{(bool vector number of cells)
Whether observation is zero.}

\item{scaNCells}{(scalar)
Number of cells (auxillary parameter, this is length of vecCounts).}
}
\value{
NULL
This will be: scaLogLik (scalar) Likelihood under zero-inflated
negative binomial model.
}
\description{
NOT YET SUPPORTED.
LINEAGEPULSE CODE WILL BE EXTENDED BY MODULAR FUNCTIONALITIES
AND THIS IS ONE INSTANCE OF A PLACEHOLDER USED FOR DEVELOPING.
Chooses whether to evaluate ZINB or NB model based on input.
}
\author{
David Sebastian Fischer
}
