% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_evalLogLikZINB.R
\name{evalLogLikMatrix}
\alias{evalLogLikMatrix}
\title{Wrapper for log likelihood of zero-inflated negative binomial model
for a matrix of counts (parallelised).}
\usage{
evalLogLikMatrix(matCounts, lsMuModel, lsDispModel, lsDropModel,
  matWeights = NULL, boolConstModelOnMMLL = FALSE)
}
\arguments{
\item{matCounts}{(count matrix genes x cells)
Observed read counts, not observed are NA.}

\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{lsDispModel}{(list)
Object containing description of gene-wise dispersion parameter models.}

\item{lsDropModel}{(list)
Object containing description of cell-wise drop-out parameter models.}

\item{matWeights}{(numeric matrix cells x mixtures) [Default NULL]
Assignments of cells to mixtures (for strMuModel="MM").}

\item{boolConstModelOnMMLL}{(bool) [Default FALSE]
Whether to evaluate constant gene-wise mean model on mixture model likelihood.
This is necessary for numeric stability when testing whether mixture model
fits are worse than constant fits.}
}
\value{
vecLogLik (vector length number of genes) 
Loglikelihood of each gene under zero-inflated negative binomial model.
}
\description{
This likelihood function is a wrapper computes loglikelihood
of entire data set by parallelising loglikelihood computation
over genes.
}
\seealso{
Called directly by \code{fitZINB} to track
convergence of estimation iteration on entire data set.
}
\author{
David Sebastian Fischer
}
