% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_fitDropout.R
\name{evalLogLikPiZINB_ManyCells_comp}
\alias{evalLogLikPiZINB_ManyCells_comp}
\title{Compiled function: evalLogLikPiZINB}
\usage{
evalLogLikPiZINB_ManyCells_comp(vecTheta, matCounts, lsMuModel, lsDispModel,
  lsDropModel)
}
\arguments{
\item{vecTheta:}{(numeric vector length linear model) 
Parameter estimates for logit linear model for drop-out rate.}

\item{vecCounts}{(count vector number of genes)
Observed read counts, not observed are NA.}

\item{vecMu:}{(vector number of cells) Negative binomial
mean parameter estimate.}

\item{vecNormConst:}{(numeric vector number of cells) 
Model scaling factors, one per cell.}

\item{vecDisp:}{(vector number of cells) Negative binomial
dispersion parameter estimate.}

\item{matPiPredictors:}{(matrix genes x predictors) Predictors of
the drop-out rate in the linear model. Minimum are a constant
offset and log of the negative binomial mean parameter. 
Other gene-specific predictors can be added.}

\item{vecboolNotZero:}{(bool vector number of cells)
Whether observation is larger than zero.}

\item{vecboolZero:}{(bool vector number of cells)
Whether observation is zero.}
}
\value{
scaLogLik: (scalar) Value of cost function:
   zero-inflated negative binomial likelihood.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalLogLikPiZINB}.
}
\seealso{
\link{evalLogLikPiZINB}
}
\author{
David Sebastian Fischer
}
