% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_fitZINB_fitDropout.R
\name{evalLogLikPiZINB_SingleCell_comp}
\alias{evalLogLikPiZINB_SingleCell_comp}
\title{Cost function zero-inflated negative binomial model for drop-out fitting}
\usage{
evalLogLikPiZINB_SingleCell_comp(vecTheta, vecCounts, vecMu, scaNormConst,
  vecDisp, matPiAllPredictors, strDropModel, vecidxNotZero, vecidxZero)
}
\arguments{
\item{vecTheta}{(numeric vector length linear model) 
Parameter estimates for logit linear model for drop-out rate.}

\item{vecCounts}{(count vector number of genes)
Observed read counts, not observed are NA.}

\item{vecMu}{(vector number of cells) Negative binomial
mean parameter estimate.}

\item{scaNormConst}{(scalar) 
Model scaling factors, one per cell.}

\item{vecDisp}{(vector number of cells) Negative binomial
dispersion parameter estimate.}

\item{matPiAllPredictors}{(matrix genes x predictors) Predictors of
the drop-out rate in the linear model. Minimum are a constant
offset and log of the negative binomial mean parameter. 
Other gene-specific predictors can be added.}

\item{strDropModel}{(str) {"logistic_ofMu", "logistic"}
[Default "logistic_ofMu"] Definition of drop-out model.
"logistic_ofMu" - include the fitted mean in the linear model
of the drop-out rate and use offset and matPiConstPredictors.
"logistic" - only use offset and matPiConstPredictors.}

\item{vecidxNotZero}{(bool vector number of cells)
Whether observation is larger than zero.}

\item{vecidxZero}{(bool vector number of cells)
Whether observation is zero.}

\item{strDropFitGroup}{(str) {"PerCell", "AllCells"}}
}
\value{
scaLogLik (scalar) Value of cost function:
zero-inflated negative binomial likelihood.
}
\description{
Refer to evalLogLikPiZINB_SingleCell().
}
\seealso{
Compiled version of \code{evalLogLikPiZINB_SingleCell}.
}
\author{
David Sebastian Fischer
}
