% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_getFits.R
\name{getNormData}
\alias{getNormData}
\title{Return depth and batch corrected data}
\usage{
getNormData(matCounts, lsMuModel, vecGeneIDs, boolDepth = TRUE,
  boolBatch = TRUE)
}
\arguments{
\item{matCounts}{(numeric matrix genes x cells)
Count data.}

\item{lsMuModel}{(list) Mean parameter model parameters.}

\item{vecGeneIDs}{(vector of strings) 
Gene IDs for which mean model fits are to be extracted.}

\item{boolDepth}{(bool) [Default TRUE] Whether to normalize for sequencing depth.}

\item{boolBatch}{(bool) [Default TRUE] Whether to normalize for batch.}
}
\value{
(numeric matrix genes x cells)
Input data normalized by library size factors (optional) and
by inferred batch factors (optional).
}
\description{
The data normalisation is based on the model normalisation used by and inferred
by LineagePulse, e.g. for data visualisation.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 20,
    scaNConst = 2,
    scaNLin = 2,
    scaNImp = 2,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 6),
    vecGeneWiseDropoutRates = rep(0.1, 6))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
# Get batch correction on alternative model:
# Use H1 model fits.
matNormData <- getNormData(
     matCounts = lsSimulatedData$counts,
     lsMuModel = objLP@lsMuModelH1,
     vecGeneIDs = rownames(lsSimulatedData$counts)[1],
     boolDepth = TRUE, boolBatch = TRUE)

}
\seealso{
Called by \code{fitZINB}. Can be called by user.
}
\author{
David Sebastian Fischer
}
