% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_initialiseImpulseParameters.R
\name{initialiseImpulseParameters}
\alias{initialiseImpulseParameters}
\title{Estimate impulse model parameter initialisations}
\usage{
initialiseImpulseParameters(vecCounts, lsMuModelGlobal, vecMu, vecDisp,
  vecDrop = NULL)
}
\arguments{
\item{vecCounts}{(count vector number of samples) Count data.}

\item{lsMuModelGlobal}{(list)
Object containing meta-data of gene-wise mean parameter models.}

\item{vecMu}{(vector number of cells) Negative binomial
mean parameter estimates.}

\item{vecDisp}{(vector number of cells) Negative binomial
dispersion parameter estimates.}

\item{vecDrop}{(vector number of cells) [Default NULL]
Dropout parameter estimates. Are set to a vector of zeros
if not supplied (i.e. if impulse model is initialised under
negative binomial noise model).}
}
\value{
list (length 2)
\itemize{
    \item peak: (numeric vector number of impulse
model parameters) Impulse model parameter initialisation 
corresponding to a peak.
    \item valley: (numeric vector number of impulse
model parameters) Impulse model parameter initialisation 
corresponding to a valley.
}
}
\description{
The initialisations reflect intuitive parameter choices corresponding
to a peak and to a valley model and are based on proximity binned data.
}
\seealso{
Called by impulse model fitting wrappers:
\code{fitImpulseZINB}.
}
\author{
David Sebastian Fischer
}
