% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\docType{class}
\name{LineagePulseObject-class}
\alias{LineagePulseObject-class}
\title{Container class for LineagePulse output}
\description{
LineagePulse output and intermediate results such as model fits.
}
\section{Slots}{

\describe{
\item{\code{dfAnnotationProc}}{(data frame cells x meta characteristics)
Annotation table which contains meta data on cells.}

\item{\code{dfResults}}{(data frame samples x reported characteristics) 
Summary of fitting procedure and 
differential expression results for each gene.}

\item{\code{lsDispModelH1}}{(list)
Object containing description of 
gene-wise dispersion parameter models of H1.}

\item{\code{lsDispModelH0}}{(list)
Object containing description of 
gene-wise dispersion parameter models of H0.}

\item{\code{lsDispModelConst}}{(list)
Object containing description of gene-wise dispersion parameter models of
constant model (necessary if H0 is not constant (mixture models)).}

\item{\code{lsDispModelH1_NB}}{(list)
Object containing description of 
gene-wise dispersion parameter models of H1 with NB noise model.}

\item{\code{lsDispModelH0_NB}}{(list)
Object containing description of 
gene-wise dispersion parameter models of H0 with NB noise model.}

\item{\code{lsDropModel}}{(list)}

\item{\code{lsMuModelH0}}{(list)
Object containing description of gene-wise mean parameter models of H0.}

\item{\code{lsMuModelH1}}{(list)
Object containing description of gene-wise mean parameter models of H1.}

\item{\code{lsMuModelConst}}{(list)
Object containing description of gene-wise means parameter models of
constant model (necessary if H0 is not constant (mixture models)).}

\item{\code{lsMuModelH0_NB}}{(list)
Object containing description of gene-wise mean parameter models of H0
with NB noise model.}

\item{\code{lsMuModelH1_NB}}{(list)
Object containing description of gene-wise mean parameter models of H1.
with NB noise model.}

\item{\code{lsFitConvergence}}{(list)
Fitting reporter summary.}

\item{\code{matCountsProc}}{(count matrix genes x cells)
Sparse matrix of counts, non-observed values are NA.}

\item{\code{matWeights}}{(numeric matrix cells x mixtures)
Assignments of cells to mixtures.}

\item{\code{scaDFSplinesDisp}}{(scalar) 
If strDispModelFull=="splines" or strDispModelRed=="splines", 
the degrees of freedom of the natural
cubic spline to be used as a dispersion parameter model.}

\item{\code{scaDFSplinesMu}}{(scalar) 
If strMuModel=="splines", the degrees of freedom of the natural
cubic spline to be used as a mean parameter model.}

\item{\code{strReport}}{(str) LineagePulse stdout report (log).}

\item{\code{vecAllGenes}}{(vector of strings) 
All genes originally supplied, including all zero genes.}

\item{\code{vecConfoundersDisp}}{(vector of strings number of confounders on  dispersion)
Confounders to correct for in dispersion batch
correction model, are subset of column names of
dfAnnotation which describe condounding variables.}

\item{\code{vecConfoundersMu}}{(vector of strings number of confounders on  mean)
Confounders to correct for in mu batch
correction model, are subset of column names of
dfAnnotation which describe condounding variables.}

\item{\code{scaOmega}}{(scalar)
Regularisation parameter on entropy penalty on mixture weights.
Needed for mixture model mode only.}

\item{\code{boolFixedPopulations}}{Whether a population of cells is to be fixed to be 
assigned to a subset of centroids (RSA scenario).
Needed for mixture model mode only.}

\item{\code{vecNCentroidsPerPop}}{Number of centroids per population (RSA scenario).
Needed for mixture model mode only.}

\item{\code{vecH0Pop}}{(integer vector)
Needed for mixture model mode only.}

\item{\code{vecNormConst}}{(numeric vector length number of cells)
Normalisation constants (library size factors) for each cell.}

\item{\code{strVersion}}{(str) Version of LineagePulse that was run.}
}}

\author{
David Sebastian Fischer
}
