% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\name{LPsetters}
\alias{LPsetters}
\alias{`dfAnnotationProc<-`}
\alias{`dfResults<-`}
\alias{`lsMuModelH0<-`}
\alias{`lsMuModelH1<-`}
\alias{`lsMuModelConst<-`}
\alias{`lsMuModelH0_NB<-`}
\alias{`lsMuModelH1_NB<-`}
\alias{`lsDispModelH0<-`}
\alias{`lsDispModelH1<-`}
\alias{`lsDispModelConst<-`}
\alias{`lsDispModelH0_NB<-`}
\alias{`lsDispModelH1_NB<-`}
\alias{`lsDropModel<-`}
\alias{`lsFitConvergence<-`}
\alias{`matCountsProc<-`}
\alias{`matWeights<-`}
\alias{`scaDFSplinesDisp<-`}
\alias{`scaDFSplinesMu<-`}
\alias{`strReport<-`}
\alias{`vecAllGenes<-`}
\alias{`vecConfoundersDisp<-`}
\alias{`vecConfoundersMu<-`}
\alias{`scaOmega<-`}
\alias{`boolFixedPopulation<-`}
\alias{`vecH0Pop<-`}
\alias{`vecNormConst<-`}
\alias{`strVersion<-`}
\alias{dfAnnotationProc<-}
\alias{dfResults<-}
\alias{lsMuModelH0<-}
\alias{lsMuModelH1<-}
\alias{lsMuModelConst<-}
\alias{lsMuModelH0_NB<-}
\alias{lsMuModelH1_NB<-}
\alias{lsDispModelH0<-}
\alias{lsDispModelH1<-}
\alias{lsDispModelConst<-}
\alias{lsDispModelH0_NB<-}
\alias{lsDispModelH1_NB<-}
\alias{lsDropModel<-}
\alias{lsFitConvergence<-}
\alias{matCountsProc<-}
\alias{matWeights<-}
\alias{scaDFSplinesDisp<-}
\alias{scaDFSplinesMu<-}
\alias{strReport<-}
\alias{vecAllGenes<-}
\alias{vecConfoundersDisp<-}
\alias{vecConfoundersMu<-}
\alias{scaOmega<-}
\alias{boolFixedPopulation<-}
\alias{vecH0Pop<-}
\alias{vecNormConst<-}
\alias{strVersion<-}
\title{LineagePulseObject setters}
\usage{
dfAnnotationProc(objLP) <- value

dfResults(objLP) <- value

lsMuModelH0(objLP) <- value

lsMuModelH1(objLP) <- value

lsMuModelConst(objLP) <- value

lsMuModelH0_NB(objLP) <- value

lsMuModelH1_NB(objLP) <- value

lsDispModelH0(objLP) <- value

lsDispModelH1(objLP) <- value

lsDispModelConst(objLP) <- value

lsDispModelH0_NB(objLP) <- value

lsDispModelH1_NB(objLP) <- value

lsDropModel(objLP) <- value

lsFitConvergence(objLP) <- value

matCountsProc(objLP) <- value

matWeights(objLP) <- value

scaDFSplinesDisp(objLP) <- value

scaDFSplinesMu(objLP) <- value

strReport(objLP) <- value

vecAllGenes(objLP) <- value

vecConfoundersDisp(objLP) <- value

vecConfoundersMu(objLP) <- value

scaOmega(objLP) <- value

boolFixedPopulation(objLP) <- value

vecH0Pop(objLP) <- value

vecNormConst(objLP) <- value

strVersion(objLP) <- value
}
\arguments{
\item{objLP}{(LineagePulse-Object)
A LineagePulse output object to write into.}

\item{value}{(str) The value to be set.}
}
\value{
The LineagePulse output object with updated entry.
}
\description{
Setters for LineagePulse-Object. Set the entry defined by the function
name to in objLP to value.
}
\author{
David Sebastian Fischer
}
