% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_plotGene.R
\name{plotCellDensity}
\alias{plotCellDensity}
\title{Plot density of cells in pseudotime}
\usage{
plotCellDensity(objLP, bwDensity = NULL)
}
\arguments{
\item{objLP}{(LineagePulseObject) LineagePulseObject to base plot on.}

\item{bwDensity}{(scalar) [Default NULL]
Bandwidth of kernel density estimate to be used. Chosen
automatically if NULL (recommended).}
}
\value{
gplotGene: (ggplot object)
ggplot2 kernel density estimator plot.
}
\description{
Uses kernel density estimate.
}
\author{
David Sebastian Fischer
}
