% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_runDEAnalysis.R
\name{runDEAnalysis}
\alias{runDEAnalysis}
\title{Differential expression analysis}
\usage{
runDEAnalysis(objLP)
}
\arguments{
\item{:}{objLP (LineagePulseObject)
LineagePulseObject with fitted null and alternative models.}
}
\value{
objLP (LineagePulseObject)
LineagePulseObject with analysis summary table (dfResults)
added.
\itemize{
\item Gene: Gene ID.
\item p: P-value for differential expression.
\item padj: Benjamini-Hochberg false-discovery rate corrected p-value
for differential expression analysis.
\item loglik_full: Loglikelihood of full model.
\item loglik_red: Loglikelihood of reduced model.
\item df_full: Degrees of freedom of full model.
\item df_red: Degrees of freedom of reduced model
\item mean: Inferred mean parameter of constant model of first batch.
From combined samples in case-ctrl. 
\item allZero (bool) Whether there were no observed non-zero observations of this gene.
If TRUE, fitting and DE analsysis were skipped and entry is NA.
}
Entries only present in case-only DE analysis:
\itemize{
\item converge_impulse: Convergence status of optim for 
impulse model fit (full model).
\item converge_const: Convergence status of optim for 
constant model fit (reduced model).
}
Entries only present in mixture model DE analysis:
\itemize{
\item converge_mixture: Convergence status of optim for 
mixture model fit (full model).
\item converge_const: Convergence status of optim for 
constant model fit (reduced model).
}
}
\description{
Performs differential expression analysis based on previously estimated 
null and alternative models.
(I) Compute loglikelihood of data under null H0 and alternative H1 model.
(II) Differential expression analysis as loglikelihood ratio test.
}
\seealso{
Called by \code{runLineagePulse}.
}
\author{
David Sebastian Fischer
}
