% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_runDEAnalysis.R
\name{runDEAnalysis}
\alias{runDEAnalysis}
\title{Differential expression analysis}
\usage{
runDEAnalysis(objLP)
}
\arguments{
\item{objLP}{(LineagePulseObject)
LineagePulseObject with fitted null and alternative models.}
}
\value{
objLP (LineagePulseObject)
LineagePulseObject with analysis summary table (dfResults)
added.
\itemize{
\item Gene: Gene ID.
\item p: P-value for differential expression with ZINB noise.
\item mean: Inferred mean parameter of constant model of first batch.
\item padj: Benjamini-Hochberg false-discovery rate corrected p-value
for differential expression analysis with NB noise.
\item p_nb: P-value for differential expression with ZINB noise.
\item padj_nb: Benjamini-Hochberg false-discovery rate corrected p-value
for differential expression analysis with NB noise.
\item loglik_full_zinb: Loglikelihood of full model with ZINB noise.
\item loglik_red_zinb: Loglikelihood of reduced model with ZINB noise.
\item loglik_full_nb: Loglikelihood of full model with NB noise.
\item loglik_red_nb: Loglikelihood of reduced model with NB noise.
\item df_full: Degrees of freedom of full model.
\item df_red: Degrees of freedom of reduced model
\item allZero (bool) Whether there were no observed non-zero observations of this gene.
If TRUE, fitting and DE analsysis were skipped and entry is NA.
}
}
\description{
Performs differential expression analysis based on previously estimated 
null and alternative models.
(I) Compute loglikelihood of data under null H0 and alternative H1 model.
(II) Differential expression analysis as loglikelihood ratio test.
}
\seealso{
Called by \code{runLineagePulse}.
}
\author{
David Sebastian Fischer
}
