% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_sortGeneTrajectories.R
\name{sortGeneTrajectories}
\alias{sortGeneTrajectories}
\title{Cluster expression mean trajectories}
\usage{
sortGeneTrajectories(vecIDs, lsMuModel, dirHeatmap = NULL)
}
\arguments{
\item{vecIDs}{(vector of strings)
Names of genes to cluster.}

\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{dirHeatmap}{(str directory) [Default NULL]
Directory to which heatmap is saved to. Return heatmap object if NULL.}
}
\value{
list (length 3)
If dirHeatmap is not NULL, only vecSortedGenes is returned and the two
heatmaps are printed to pdfs in the directory dirHeatmap.
vecSortedGenes: (string vector number of IDs)
hmGeneSorted: genes sorted by peak time in pseudotime
hmGeneClusters: genes sorted by clustering
}
\description{
Sorts inferred gene trajectories by peak time in pseudotime. Optional: Can
create a heatmap of the gene trajectories sorted according to peak time.
The heatmap is based on z-scores.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 100,
    scaNConst = 10,
    scaNLin = 10,
    scaNImp = 10,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 30),
    vecGeneWiseDropoutRates = rep(0.1, 30))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
lsHeatmaps <- sortGeneTrajectories(
    vecIDs = objLP$dfResults[which(objLP$dfResults$adj.p < 0.01),]$gene,
    lsMuModel = lsMuModelH1(objLP),
    dirHeatmap = NULL))
#print(lsHeatmaps$hmGeneSorted)

}
\seealso{
Called by \code{LineagePulse_main} or by user.
}
\author{
David Sebastian Fischer
}
