% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_sortGeneTrajectories.R
\name{sortGeneTrajectories}
\alias{sortGeneTrajectories}
\title{Cluster expression mean trajectories}
\usage{
sortGeneTrajectories(vecIDs, lsMuModel, dirHeatmap)
}
\arguments{
\item{lsMuModel}{(list)
Object containing description of gene-wise mean parameter models.}

\item{dirHeatmap}{(str directory) [Default NULL]
Directory to which heatmap is saved to. Not 
created if NULL. Need to have lsMuModel$lsMuModelGlobal$vecPseudotime
defined.}

\item{vecIDs:}{(vector of strings)
Names of genes to cluster.}
}
\value{
vecSortedGenes (string vector number of IDs)
IDs sorted by peak time in pseudotime.
}
\description{
Sorts inferred gene trajectories by peak time in pseudotime. Optional: Can
create a heatmap of the gene trajectories sorted according to peak time.
The heatmap is based on z-scores.
}
\examples{
lsSimulatedData <- simulateContinuousDataSet(
    scaNCells = 100,
    scaNConst = 10,
    scaNLin = 10,
    scaNImp = 10,
    scaMumax = 100,
    scaSDMuAmplitude = 3,
    vecNormConstExternal=NULL,
    vecDispExternal=rep(20, 30),
    vecGeneWiseDropoutRates = rep(0.1, 30))
objLP <- runLineagePulse(
    counts = lsSimulatedData$counts,
    dfAnnotation = lsSimulatedData$annot,
    strMuModel = "impulse")
#sortGeneTrajectories(
#    vecIDs = rownames(objLP$objLP@lsMuModelH1$matMuModel),
#    lsMuModel = objLP@lsMuModelH1,
#    dirHeatmap = "your_target_directory"))

}
\seealso{
Called by \code{LineagePulse_main} or by user.
}
\author{
David Sebastian Fischer
}
