% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcLineagePulse_classLineagePulseObject.R
\name{writeReport}
\alias{writeReport}
\title{Print LineagePulse report}
\usage{
writeReport(object, file = "")
}
\arguments{
\item{object}{(LineagePulseObject) Output object of LineagePulse.}

\item{file}{(file) [DEFAULT ""] File to print report to. Default is stdout.}
}
\description{
Print LineagePulse report string to .txt file or stdout.
}
